transactions: spend only confirmed outputs
Spending unconfirmed outputs is not secure, as the output might never
confirm.

Exception: we allow to spend unconfirmed outputs created by our own
wallet. It is a tradeoff to improve the user experience where the
available balance is influenced by invisible unconfirmed change
outputs. The downside is with tx fee priorization when spending the
unconfirmed output, which is hopefully a mostly invisible edge case.

Example: wallet has one 10 BTC output. Spend 1 BTC, 9 BTC are
change. The user would expect to have 9 BTC available after, but since
the change is unconfirmed, it would jump to 0.