frontend: Move self-signed certificate to appropriate directory.
A self-signed certificate is generated when the application starts.
Previously, this certificate was put into the config/ directory as server.pem. Now it's
generated in config/certificates/frontend/server.pem to avoid ambiguity.

Also: Added a method that cleans up test certificates when the test is done.

Also: Added additional packages to build the qt app in docker_install.sh.