scripts: add script to run bitcoind+electrumx in regtest
Run the script, then use `make servewallet-regtest` and `bitcoin-cli`
to interact with it.

Use regtest for functional tests.

Start the regtest backend with ./scripts/run_regtest.sh. Once it is
running, the backend functional tests will automatically pick it up
and use it.

Use ./scripts/run_functional_tests.sh to run the functional tests of
the backend using the regtest server.