client: take minimum relay fee if electrumX return fee=-1, revisited
Problem: rpc callbacks are handled in a single thread. RelayFee() was
blocking, waiting for the response of a future rpc callback, but since
it was called inside one of those callbacks, there was a deadlock.

Easy fix would be to run all those callbacks in goroutines, but it's
not nice to run goroutines if you don't know how many there will be.

The fix in this patch makes RelayFee() async as well.