pairing: Adapted pairing process
The pairing mechanism between mobile app and bitbox changed, therefore the
godbb app must adapt to the new mechanism and commands.

The mobile app and bitbox commit to a public key before exchanging them.
Thus, the hash of the public keys are exchanged first.
Furthermore, the challenge command that triggers the blinking of the value of
the shared secret is now triggered manually and must therefore also be forwarded
to the Bitbox.