maketx: correctly compute tx sizes and fees
Before, p2pkh was assumed for all inputs and outputs/change, which
distorts the tx size and fee estimation. Different address types have
different sizes in the output and input. Also, the concept of tx
weight (virtual size) was ignored, losing the segwit fee discount.

Now, the tx sizes are computed correctly and as a result, fees are
generally lower.