backend: partial functional tests fix
Problem: a new header modifies the state via address history pushes
from electrumx, and by new header pushes from electrumx (for tx
verification). Both can happen at the same time or in sequence, but
both use the same EventSynced event after the state change is
done. This breaks functional tests, as the sequence of the events is
not clear.