qt,server: stop using self signed certs
Now, the api is bridged directly without a http server. The frontend
is still served via http. Intercepting this or visiting it outside the
app won't do much harm, as it's only known static files.

In the future, we might be able to serve this also without a server.

The code to generate self signed certs has been moved for future
reference.