frontend: disable 2FA button unless paired
This change simply prevents enabling 2FA from the web UI
if no phone is paired with the device, as reported by the backend.

It remains to be seen and verified whether the 2FA can be enabled
directy by sending "lock" command to the device
or backend which propagates the command to device.