backend/rates: fix race condition and redundant notifications
Rates were observed when coins were initialized, which is after
unlock. The first rates notification happens immediately when the
backend is initalized, so the frontend misses the first notification,
and waits a full minute for the first one arrives.

Also, for each coin, one notification was made, but there is a single
global observer/rates variable for all coins in the frontend, so it
does redundant work.

Now, the rates are removed from coins and are part of the backend, to
match how they are used in the frontend. Can move it back if we ever
do per-coin fetches, but that seems unlikely.