Makefile: locize push should also add/remove keys in every language
Say someone adds a key to English, and in the same go adds it to some
other languages they know. Locize push would lose the changes to other
languages (something that should be avoided, but if it happens, we
don't want the work to be lost). This flag fixes that.