Add BitBoxBase Support for Demonstrations
This is a barebones implementation to demonstrate connection with the BitBox base middleware.

MDNS scans for bitboxbase tcp service entries in the background. If a service entry is found, TryMakeNewBase() is called, which creates a connection with the base.
Once a connection to the middleware has been established with the ip and port entered in the frontend, the backend creates a new bitboxBase and bitboxBaseID, and finally emits an event that subsequently spawns a new base component in the frontend.

This connectBase component in the frontend then subscibes to the blockinfo endpoint of the middleware, which has been implemented for demo purposes. The backend is also capable of adding and removing connections to multiple bitbox base nodes, the frontend provides a 'delete' button for each
base component to showcase this. If the entry is removed in the frontend, mdns does not try to
re-register it again -> needs to be blacklisted as well. It is acceptable for now to keep this blacklist in RAM.
The BitBoxBaseIDs are sorted by time created before being passed to the frontend.