backend/coins: CSV export - spread multitx on multiple rows
Easier to do accounting/summing etc. if transactions containing
multiple transfers (multiple addresses sent to or received on) are
spread on multiple rows. Before, the sum of all was collapsed in one
row.

The tx fee is only written in the first row if the same tx id appears
in multiple rows, to avoid double booking.