Add option to run devmode servewallet on prod
The devmode electrum servers do not persist any change to their
configuration when the app is restarted. To allow the BitBox Base to
persist its config on restart in the future, a devservers flag is added.
It is set by default to true when running servewallet. To test on
production servers, the Makefile command `make servewallet-prodservers` is
added.