Allow connection to bitbox-base electrs
This commit adds support to connect with electrs running testnet or
mainnet on the BitBox Base. A new postConnectElectrumHandler was
created for this purpose. It takes a bitboxBaseID, from that gets
the connected bitbox Bases'IP and initiates a connection with the
base.

The electrs connection is handled by the bitboxbase package. To
facilitate this the DownloadCert and CheckElectrumServer functions are
moved from the backend to the electrum package.

The electrs port and network type (mainnet/testnet) are fetched with the
GetEnv get request.

Once connected, other active backends will be unset, such that
only bitcoin accounts will be available.