base-detection: use manager to only detect new bases
The user story is that when the app discovers a new base on the
network, the user should manually select to connect to it.
For this, the app only needs the IP of the discovered base.
The backend should only create the necessary base interface and
connection once the user decides to connect and until then only keep
track of the minimal information about the discovered bases. The IP
for connecting and the hostname to display to the user