bitbox-base: implement minimal UI for detecting and connecting
This PR implements base mdns detection and connection based on the
initialization flowchart and wireframing diagrams.
When clicking on the bitbox base menu, the user can see all the bases
detected on the local network with mDNS and choose which base to
connect to. Pressing connect connects to the selected base's
middleware.
If no bases are detected, the user still has the option to connect
manually by entering the base IP if the mDNS discovery fails