Frontend/Backend BitBoxBase: Initialization Wizard
This commit adds a basic initialization wizard to the frontend. It
tracks the current position in the wizzard with a status string from the
bitboxbase. Changes to the status are communicated to the frontend by
the means of websocket notifications.

A connected base now gets its own sidebar entry. According to the
wireframe designs, this should be changed in the future, where a view
of a singular base replaces the bitboxbaseconnect view.

The wizard currently tracks the progress during the noise pairing with
the base. It starts at 'unpaired' and moves to 'initialized' if
successful, or 'pairingFailed' if not.

The transitions are triggered by events, either signifying that the
noise channel hash state has changed, or that the status has changed.

The noise pairing verification had to be tweaked a bit in order for this commit
to work. The BitBox Base middleware server now expects a message even if
no verification is required. This makes sure that no rpc
notification is emitted by the middleware while the wallet app is still
listening for the verification response.