BitBox Base: Use events to trigger actions in frontend
This commit introduces a way for events in the BitBox Base middleware to
reach the frontend and then trigger a response there.

If new data becomes available in the middleware, the middleware sends a
notification wallet app over the noise encrypted websocket channel. The
notification is caught in the rpcclient's parseMessage function. The
rpcclient calls a callback function from the bitboxbase that in turn
triggers an event for the frontend.

Once the notification has reached the frontend, it is interpreted in a
switch statement. In the case of a middleware notification, it sends an
http request to the bitboxbase handlers, that through additional
validation in the bitboxbase and rpcclient package triggers an rpc call.
The return value of the rpc call can then be post-processed by the
bitboxbase package and then returned to the handler, which sends it back
to the frontend as a response to the original http request.

Having events propagate from the middleware directly to the frontend has
a few advantages. Notified resources are only fetched from the
middleware if the frontend component that needs them is currently mounted.
It also reduces the amount of time-series storage that the app backend
has to do in order to pre-cache the notified values.