Backend: Add Tor Socks Proxy Support
To allow app users to route all the traffic of the BitBox app through
tor, the backend is now able to read a socks5 proxy address and if
activated, will route all web traffic through this proxy.

All web endpoints need to use the socksproxy package to manage their
connections to ensure that no clearnet traffic is sent anymore.