BitBoxBase: Add Authentication with JWT
All rpc calls now need to be authenticated with a token. A new token is
received by calling the UserAuthenticate rpc method with the users's
password and username.

This commit also re-works the setup wizzard for the BitBox Base to make
sure that the authentication is run correctly and that the wizzard is
not left in a state from which it cannot recover.