frontend: lazy load qrreader
The @zxing/library that is used for the BrowserQRCodeReader was
bundled in the main JavaScript bundle and is quite big.

This change imports the BrowserQRCodeReader asynchroniously when
the send view is mounted. This decreases the main bundle size.
Ideally we would load the reader only when the user clicks on
the qr scanner icon, but we need the library already on send
to check if the scanner is available at all.