ci: make golangci-lint great again
This resolves the following error

    + golangci-lint run
    level=error msg="Running error: context loading failed:
    no go files to analyze"
    Makefile:79: recipe for target 'ci' failed

I believe this is because of a recent golangci-lint upgrade.
It seems to be running with the default go modules support now,
which in turn tries downloading all the package dependencies.

Unfortunately, one package dep was added to go.sum at the time
where package checksums were computed incorrectly in certain conditions.
More on that can be found in golang/go issue 29278.

This went unnoticed for quite some time because vendored dir is used
instead of go.mod. Since the upgrade of golangci-lint, it started
using go.mod.

This change updates go.mod checksum, even though it isn't used,
and makes golangci-lint use vendored dir as it should.