bbBase: persist Bases in backend and auto-reconnect
Because:
If a Base rebooted, shut down or the connection was lost, it was
automatically removed from the backend and the user would have to
reconnect to it manually.

This commit:
- Introduces three new statuses: DisconnectTypeReboot,
DisconnectTypeShutdown, and DisconnectTypeDisconnect to distinguish
between different types of disconnects and take appropriate
action.
- E.g., DisconnectTypeReboot will trigger an
`attemptReconnectLoop()` go routine to reconnect to the Base once the
middleware is reachable again.
- Moves the Ping() function from the `rpcclient` to `bitboxbase`,
because the RPCClient needs to be re-created when reconnecting and we
need to ping the middleware from the App without relying on the
RPCClient which is now `nil`