base: implement user-facing error messages based on error codes
When an RPC to the BitBoxBase fails, the Base returns a predefined
error code from errorcodes.go, this commit maps these error codes
to user facing messages with the fallback value of
```"UNEXPECTED_ERROR": "An unexpected error occurred. Please see
application log for additional details."```
in the case the Base returns an unknown or empty error code when an
error occurs.