backend/btc/transactions: download txs sequentially
The previous mechanism could download and process txs asynchronously,
but with this, an address could not be processed using a single
database transaction. This meant that if the app was closed or crashed
before a tx could be downloaded, the db could be in an inconsistent
state, and show the wrong balance.

This commit changes this so that tx downloading is blocking, and one
address is processed in a single database transaction. This means that
on a sudden app close, the address will be reprocessed, starting from
a consistent state.

This might mean potentially slower syncing time if some txs have a
high fetch latency, but in my experiments, there was no slowdown.

Side-benefit: the transactions syncing code is a lot simpler now.