backend/btc/electrum: use TLS setting in Electrum connection check
The GUI currently only supports adding TLS enabled servers, so the
check also assumed TLs to be active. If TLS is actually enabled is a
flag in the app config file and can be false (currently not possible
in the app UI). The check now works also if flag was manually changed
in the config file.