backend/keystore: drop multisig arg as well
It was only used for the bitbox01, where multisig is supported in a
demo but address verifcation is not.

Simple solution is to disable multisig in the bitbox01. If we need to
revive the multisig demo, it can be with the bitbox02 in general
devices that support the address verification.