backend/btc: bundle address chains and signing config in one struct
In preparation for unifying account types into one account.

I already make this a slice, but with only one element, so it behaves
the same as before.

When unifying e.g. Bitcoin Legacy, Segwit and Native Segwit accounts,
there would be one entry per such signing configuration.

Each accesss is marked with:

// TODO unified-accounts

with each following commit fixing one of those. This hopefully results
in smaller and easier to review commits.