backend/btc/maketx: simplify loop
The initial target fee is calculated to skip a loop iteration in the
rare case that the first input is large enough to cover the output,
but not large enough to also cover the fee.

The initial fee assumes that the first input has a certain script
configuration. Since we want to allow mixed inputs, this estimation
does not work anymore.

Removing the initial fee estimation does not change the output.