backend/config,frontend/settings: toggle Bitcoin, Litecoin
Not the individual subaccount types.

For combined accounts ("Bitcoin", "Litecoin"), it just toggles the
account.

For split accounts, now either all account types are shown (Bitcoin:
bech32, Bitcoin, Bitcoin Legacy if supported), or none.

This is an acceptable trade-off, and only affects users with the
bitbox01 or users that will manually opt-in to split accounts.

The alternative would be a very hairy and granular settings screen,
which defeats the purpose of trying to simplify it.