backend/eth: propagate errors from EstimateGas()
When the transaction is invalid (e.g. a contract destination without
any data input), EstimateGas() returns an error.

Before, the error would be swallowed (only logged) and in the frontend
nothing would happen. Now we propagate it as is from EtherScan (it
does not provide any useful error code, just -32000 which stands for
server-error)