util/socksproxy: remove excessive logging
The "using an unproxied ..." log line does not help in any way
but it keeps repeating every time something's using it, like the
exchange rate updates. It is actually any of these and more:

    backend/update.go:
    client, err := backend.socksProxy.GetHTTPClient()

    backend/devices/bitbox/relay/request.go:
    httpClient, err := request.channel.socksProxy.GetHTTPClient()

    backend/coins/eth/etherscan/etherscan.go:
    client, err := etherScan.socksProxy.GetHTTPClient()

    backend/rates/rates.go:
    client, err := updater.socksProxy.GetHTTPClient()

    backend/backend.go:
    httpClient, err := backend.socksProxy.GetHTTPClient()

When logged, it can't be attributed to the caller.
It's a separate line and logging is typically buffered; line ordering
should not be relied upon.

Similar reasoning with when the connection is proxied.
This should be logged once, at the app startup, or by the caller of
NewSocksProxy if really necessary. This commit opts for the former.

One other part is devices/bitbox/relay but this is bb01 and the proxy
is always unused.