ci: upgrade golangci-lint to v1.27 and make it happy
Additionally, exclude some annoying and useless linters, especially in
test files.

Unfortunately, I went into the godot linter hole deep enough before it
was too late. So, here it is. On the bright side, doc comments look
more consistent now. Why a doc comment must start with the type name
but not end with the dot.

This commit also adds a separate golint-ci job as a GitHub workflow
action. This allows for nice annotations right in the Checks tab on a
PR like shown here: https://golangci-lint.run/usage/install/.