frontend/account: show number of synced addresses in the spinner
On a slow network or on Tor, the initial sync can take a very long
time (minutes). After 10 or 20 seconds without any progress
indication, the user is left guessing if it will ever finish.

Unforuntately, the number of addresses that will be synced is not
known upfront, it could be 20 or 1000 addresses, and it is a rolling
window until there are 20 (or whatever the gap limit) unused addresses
in a row.

We could show `synced X out of Y` addresses, which both numbers
creeping up, but after some discussion decided that this does not help
more than just showing the number of addresses synced. The user still
does not know when it will finish, but at least sees that there is
progress.

The additional line is only shown for >1 synced
addresses. Single-address accounts like address-watchonly or Ethereum
do not show this progress line.