frontend: click on anchor should not propagate
The A component is used for opening a link in the default browser.
If there is another click event listener on a parent of the A component,
there browser will open, but the click would also bubble up the dom.
This fix now stops the event so it should not do anything except opening
a browser.