frontend: refactor and improve coincontrol for small screen
Moved from tables to responsive design, on small screens
the addresses are shortened with ...
Moved expand icon from transactions/components to components.
Whole UTXO are is now clickable.
Rearanged UTXO item to display coin and fiat amount first.
Split outPoint into txId and txOutput, to be able to shorten
the txId and always display the txOutput.
Added 'Continue' button to the end of the dialog.