fix completely removing a note
Previously the note in detail view would either show the
newly changed note or if falsy fallback to the note passed
as prop. This was buggy and did not display the empty string
of the new note when the user removes all letters.

Because the note (state) is always copied from teh note passed
in as prop it is sufficent to just show the note state (newNote).