backend/accounts: store tx note when sending a tx
Then the note will be shown in the tx history when the tx appears
there.

Also squashed inside:

frontend/send: correctly update tx proposal when entering the note

Before, no tx proposal was made, so it was possible that there would
be no note in the active tx proposal when proceeding to
review/sign/send.