backend/btc/transactions: store tx first seen time and sort by it
Unconfirmed transactions would appear in an arbitrary order in the
app. This commit adds a time of first seen to each tx, and sorts by
that instead, so even unconfirmed transactions are ordered by age.

Same for transactions that confirmed in the same block, which are also
sorted in no particular order currently.