android: make sure device list is filtered by vendor/product ID
Before this change, any device plugged into a USB-C would be accepted
as a BitBox02 device. This results in app crash if the the device isn't
actually a BitBox02 because Go's USB manager expects only valid devices
on the backend.

This commit simply ensures that the Android native code skips unknown
devices when looking for a BitBox02 in the list.