frontend: rearange fee input, fee text and note input
Proposed fee next to the fee target didnt look great on medium and
large screen, too much information and text in the middle. Also
the proposed fee text always showed info about btc and ltc.

The fee input now shows proposed feetext below the input field,
just as text, as it is not editable anyways and does not have to
be styled like an form input. Wait time has been renamed to be
confirmation time and the info do now only display the relevant
time and block info based on the account coinCode.

Note input field is now the last field (also on small screen),
as it is optional.