frontend/send: no autoformatting of custom fee rate
The `Number()` conversion automatically cut off e.g. a zero, so it was
impossible to enter `1.01` without it being truncated back to `1` when
entering `1.0`.

Also, a default rate of `1` might not be good.

Solution: treat the fee rate as a string, like the amount, and let the
backend convert/validate.