frontend: always reset the feeperbyte when changing feetarget
Before when setting a custom fee and chaging back to a
different feetarget than custom it would still show
the custom fee in the confirm dialog.

Slightly increse the width of the confirm dialog so amounts
in ETH transactions are less likely to break on newlines.

Add icon to dropdown so it is clear that priority is a dropdown.