backend/eth: properly rate limit EtherScan
We made one EtherScan instance per ETH and ERC20 token, all of them
firing at once. This was usually not a problem, except for in the
account summary / portfolio page, where all accounts are loaded in
quick succession.

This commit moves out the rate limiting to a new httpClient reused by
all instances.