backend/rates: make historical rates settings configurable at runtime
This is done in two places: at the backend startup and whenever active
coin accounts or selected fiats list changes. In both cases, all
historical rates goroutines are shut down and then started again
but only for the selected coin/fiat pairs.

Maybe a better way would've been to shut down only those goroutines
for which no coin or fiat is active anymore, but this complicates a
simple function and with an added jitter in a follow up commit will be
unnecessary.