backend/eth: initial sync fix
An account should enforce Balance() can be safely called after
Initialize() and that the WaitSynchronized() call inside waits for it
to be actually synchronized. in BTC this is enforced as the first
counter increment happens in Initialize() (ensureAddresses()). In ETH
however, it did not. This fixes it, so that account.Balance() returns
the correct data after account.Initialize().