backend/rates: evaluate time.Now at the coingecko API call time
Before this commit, forward history update loop has never had a chance
to catch up. It always calls an update with (start, now) but by the
time it gets to fetch the data, "now" is in the past.

This gets worse as the number of active coin/fiat pairs increases.
The bigger the number, the greater the lag behind "now" is due to rate
limiting.

The change resolves the issue by evaluating "now" at the time
RateUpdater requests data from the upstream and always returns the
latest available exchange rates.