backend: reduce excessive logging
Removed some of the most offending lines which provide little useful
info and better be handled in tests.

There's more, especially when network connectivity is down, but it is
more complicated and the plan is to refactor some packages like the
jsonrpc anyway. Leaving it for later: we now rotate the logs so the
overall impact is minimal.