backend: move moonpay from account to main handlers
The moonpay widget URL needs fiat code and preferred user language.
Both are inaccessible from account handlers. They are, however,
availble in app config through the backend which the main handlers have
access to.

I'd like to refactor a lot more in the handlers space and possibly
move these handlers back to accounts but right now this seems the
quickets and easiest.