backend/config: migrate userLanguage from frontend to backend
The backend now needs to know user preferred UI language to propagate
it to the Moonpay widget.

Alternatives are to either have the frontend append the language to the
widget URL or turn opaque frontend config into a struct. Appending to
the URL form the frontend is not ideal because we might want to do some
URL signing in the future and turning frontend config into a struct
means chasing after all fields in the frontend source code. Would've
been easier from the start. Can do a larger refactoring later.