backend: unify exact match and regexp into prefixed URL matching
There's no reason why keep two separate whitelists in the SystemOpen
function. They serve the same purpose and can be easily combined into
one.

So far, there doesn't seem to be a reason to use a more onerous regexp
compared to just a string prefix. This commit in fact does the latter
and all existing patters line up nicely.

If there's really a need in the future, can always switch to a regexp
based approach.