frontend: fix backup view jumping long list
In earlier BitBoxApp versions it was possible to create new backups
of the same seed, so some old users might have long lists.
Selecting a backup from the end of the list currently causes a
weird jump in the UI and messes up scrolling.

This change fixes the jumping / scrolling when selecting a
backup of the end.

Also fixed displaying very long backup names, that they can break
onto 2 lines.