frontend: route to create and restore success screen
The success screen did not show up after restore from backup
due to the app.tsx rendering again and loosing the wizard state.

This refactors create and restore success message into its own
view. The routeing happens on keystoreAvailable, with the drawback
that it could race with the incoming accounts and keystores reload
Events.
An global frontend app state could take care about this in the
future.