backend/eth: rebroadcast transactions that seem to be lost
By experience, it seems that at least with the Etherscan backend, a
transaction that is broadcast successfully (api endpoint returns 200
OK, no error message) still might be lost and never be mined. This
lead to the ETH account being unusable, especially because follow up
transactions get a new nonce, and it also can never be mined until the
previous nonce was mined.

This is an attempt to fix this: we simply rebroadcast when we see that
the backend does not know about the transaction. It happens the first
time immediately after broadcast, and then regularly with each account
update.

For now we simply log the broadcast attempts, so we can analyse the
problem further. In the future, we might use this to mark a tx as
failed (e.g. more than 10 failed attempts), or to give the user
certain options.