backend/config: migrate node addresses from 5xxxx ports port 443
We've received a number of complaints about our nodes being
inaccessible on some systems. Most resolve by allowing an outbound
connection in a firewall or similar.
This commit tries to reduce this issue to a minimum by abusing HTTPS
and making the app connect via port 443 which these firewalls are
expected to be more lenient with.

These are the same nodes originally migrated to in commit 309ea92,
just behind port 443. On the infrastructure side, the routing is done
by matching the server name during TLS SNI handshake and
reverse-proxying TCP connection to the specific 5xxxx port on the same
host.

This of course requires client cooperation where it sends appropriate
server name in TLS SNI. Done in a separate commit.