frontend: update copy component state init, var and type naming
Moved state initialization out of constructor and use readonly
state initialization. We already do that in various places, reason
is less boilerplate and it tells TypeScript that the state key
shall not be mutated, example this.state.foo = 'bar'.
The linter should already know that this.state should be readonly
via eslint-plugin-react.

A while ago the copy component's input element has been changed
to textarea to support multiple lines, but the variable names and
types where still using input as name and type.
Changed variable names and types to use textarea, this should not
have a practical difference.