frontend: reduce routing and fix restart app routing
Currently the app and other components route in case devices
change. It is not always clear which component routes or in what
order.
The app should route when a new device is found, unplugged,
accounts are loaded, wallet created/restored. In devmode we want
to keep and reload the current view.

Changed:
- app: only route to device view if backend pushes a new deviceID
- devicesswitch: remove component routing, so app is in charge
- bitbox02: after unlock route to / and wait for accounts
- app: maybe route when backend pushes accounts
- app: removed mayberoute on componentmount, but keep mayberoute
  after fetching devices and accounts
- app: removed mayberoute in "route changed" callback

Test:
- route after unlock, get started
- route after app restart with unlocked device
- route on different views when device is unplugged
- devmode: do not route when app restarts or reloads