frontend: refactor status.tsx for preact 10
In Preact X props.children is not always an array anymore see
https://preactjs.com/guide/v10/upgrade-guide/#propschildren-is-not-always-an-array

For this reason the way status.tsx checked if it had children to
display broke and empty status messages are visible. Also various
ways of rendering message text lead to inconsistent code, as
conditionally rendering the status component or its children
with logical or ternary operator.

Added a simple hidden property that is opt-in and can be used to
hide the message and render null.

Also changed the dismiss close element from using anchor with
href=# to button, so it does not change the url#hash.