Dockerfile: update to Qt 5.15.2 and Ubuntu 18.04
Qt 5.15.2 does not support Ubuntu 16.04 anymore.

For good measure, we also upgrade linuxdeployqt.

The '-unsupported-allow-new-glibc' is needed because linuxdeployqt
tries to prevent users from building on Ubuntu 18.04 while 16.04 is
still not end-of-life, so it keeps working on Ubuntu 16.04. We
override this because 16.04 is end-of-life already.

The android base image is also upgraded from thyrlian/android-sdk:3.0
to thyrlian/android-sdk:6.0 because there is some error building the
image with the older version.

Qt5 Dockerfile: `-skip qtdocgallery` because the docs are not needed
and the there is a compilation error without it. `-skip qtquick3d`
because I stumbled across this flag somewhere and added it because it
seems unnecessary. There might be more such flags we could add in the future.