frontend/settings: validate socks proxy endpoint before persisting
An invalid endpoint, e.g. '127.0.0.1:9050XXX', was previously stored
without any checks. When starting the app, NewBackend() gets a http
client using the proxy and fails, leading to a crash due to a
panic. Even removing the panic, it is better UX to show the user the
basic errors upfront.