frontend: rewrite status component with hidden prop
Added hidden prop so the component can be told if it should render
or not and doesn't require to be wrapped in somekind of condition.

In some situations the whole status component was wrapped in a
condition, additionally status checked if its children are truthy
and rendered based on that.

Added a simple hidden property that is opt-in and can be used to
hide the message and render null.

Also changed the dismiss close element from using anchor with
href=# to button, so it does not change the url#hash.