frontend: ensure state is up to date before continuing
In send component as well as BitBox01 password and device components
the state is modified and expected to synchroniously update.
This was the cause of a few bugs in Preact X upgrade.

Changed to use the setState callback to ensure that the state
is updated before the next function continues. A hint is usually
if the next method depends on an updated .

Also removed a `// @ts-ignore` and added a check for a reference.