backend/keystore: add RootFingerprint()
And implement it for all three keystore backends we have.

Reason: we will persist accounts in the accounts.json database along
with the fingerprint of the keystore the accounts were from from. That
allows to only load the accounts belonging to the active keystores, or
in case of watchonly to know if the right keystore is connected.