backend/handlers: add /supported-coins
Will be used in the first step of adding a new account. Only the coins
supported by the keystore will be shown.

ERC20-Tokens are not included, as they will be enabled/disabled per
Ethereum account.

For now we keep it simple and assume one connected keystore. The
general solution would include querying all keystores (all connected
BitBoxes) and showing them also in the frontend in the
selection. Since the backend only allows connecting one device at a
time, we skip the effort.