backend: explictly only support one keystore
We implicitly only support one keystore (e.g. RegisterKeystore()
resets previous keystores, DeregisterKeystore() does not have a
keystore ID argument, etc). Having a keystores type handling multiple
keystores is more confusing than helpful, as most of the behavior is
not well defined, since we don't actually make use of it.

This patch makes this explicit and simplifies some code in the
process.

The accountsLock is extended to cover the keystore as well.