frontend: support deactivating account prototype
In BBApp 4.27 altcoins could be deactivated by a toggle in
the settings view. With multiaccounts support this got lost,
once an account is created it cannot be removed again.

BIP44 accounts can't technically be deleted, coins would
still be there even if the account is removed. Also remove
could be scary as a user might be affraid that coins could
be lost.

Added toggles to deactivate coins on the manage accounts
view.

TODO:
- backend api set-coin-active
- indicate or warn when deactivating an account that has coins
- add error when the last account is deactivated, at least 1
  account must be active