frontends/qt: make the libserver build work on go1.16
This resolves a CGO-related compile time errors:

    # github.com/digitalbitbox/bitbox-wallet-app/frontends/qt/server
    _cgo_export.c: In function 'serve':
    _cgo_export.c:48:3: error: expected specifier-qualifier-list before 'pushNotificationsCallback'
       pushNotificationsCallback p0;
       ^~~~~~~~~~~~~~~~~~~~~~~~~

The reason is, when a CGO-exported Go function has args matching
the type, even if it's a C type, like so:

     //export serve
    func serve(pushNotificationsCallback C.pushNotificationsCallback)

the C code generated by CGO becomes invalid due to the arg and its type
taking identical name. Here's a snippet example:

    CGO_NO_SANITIZE_THREAD
    void serve(pushNotificationsCallback pushNotificationsCallback, ...)
    {
      __SIZE_TYPE__ _cgo_ctxt = _cgo_wait_runtime_init_done();
      typedef struct {
          pushNotificationsCallback p0;
          responseCallback p1;
          notifyUserCallback p2;
          char* p3;
      } __attribute__((__packed__, __gcc_struct__)) _cgo_argtype;

Note the resulting pushNotificationsCallback.
Simply renaming the arg names resolves the issue.

Unclear when this has changed. Release notes for go1.15 and go1.16
don't seem to include anything about it.