frontend: fix bitcoin or crypto wording in the buy view
The text in the buy information uses Bitcoin or crypto depending
if a bitcoin only is used. This changed with multiaccounts, one
could for example only configure Bitcoin accounts, or use a
Bitcoin-only device.
In the past the account codes contained scrtipType information,
but with multiaccount this changed to some unique account id.

Changed buy info to detect if only Bitcoin accounts are available
and display Bitcoin or crypto. Changed buy view to display
the correct name.