frontend: hide sidebar during unlock
The sidebar on the unlock screen has not much purpose, but is
distracting, doesn't look good with pending AOPP request, and is
not very intuitive to navigate.
The sidebar is already hidden during create-new-wallet or
restore-from-backup.

The main purpose to show the sidebar is to be able to go to
settings before unlocking/installation to setup the Tor proxy.
This is for advanced users that might as well unplug the BitBox
to be able to go to the settings view.

Forced hiding the sidebar also during BitBox02 unlock.