frontend/aopp: better support for aopp banner
Made the content area grow to the full content height, instead of
specifying 100% height. That way the content area can be smaller
when there is one or more banner (app update, mobile data warning
and/or AOPP in progress banner).

Also ensure that if 2 fullscreen components are used at
the same time, i.e. AOPP user-aproval + unlock, that
the one inside the content area has a lower z-index.