frontend: add fullwidth and width props to fullscreen component
Sometimes it is prettier to present media on the whole width of
the screen, and set the width of the container. I.e. for the video
on the unlock screen.

Also left aligned the status message that is used on the unlock
screen to display device attestation error. All status messages
are left aligned, but in the fullscreen component it inherited
text align center.