backend/coins/eth: add fee targets based on EthGasStation
https://ethgasstation.info/ provides gas price estimates for different
priorities.

We use these to give the user fee options.

If the service is not available, we fall back to the previous gas
price estimation provided by the eth_gasPrice RPC call to the ETH
node (currently etherscan proxy).

FeeTargets() is currently doing all the work whenever it is called,
which should be once when the send screen is loaded and every time a
tx proposal is made (every time the user changes something in the send
screen). This could be optimized with a cache that is invalidated
regularly (or when a new block arrives).

Custom gas price option to be added later.