frontend: move steps component to device/bitbox02
The steps component used to be a step indicator so that
the user could see on-which-of-howmany steps in a flow
they currentl are. This functionality was removed somehow,
but the steps component is still used only to show/hide
the containers.

In the meantime we have a new steps component that added
the functionality again, it is currently located under
routes/account/add/components, but can be moved to the
main components once it will be reused by other modules.

Moved the legacy steps to device/bitbox02 and fixed
backups because it also depened on 1 of the styles.