frontend: use react
The current UI still uses an older verison of Preact and upgrading
to the newest Preact X has not been straightforward. A reason is
that a few workarounds had been necessary to make certain modules
work with Preact+TypeScript, i.e. react-i18next

Preact used to have less dependencies and no license issue, but
there is no real reason to keep using Preact anymore.

This is the first step of migrating from Preact to React.

Currently most components are excluded and only small POC with a
handfull of i18n related components are used in app.tsx, as well
as the LangaugeSwitch. Over the next few commits more components
will be imported and migrated.

- changed from preact@8 to react@17 using a fresh installation
  of create-react-app with `--template typescript` and `--use-npm`
- removed polyfills, recently updated webengine nativelly supports
  polyfilled JavaScript api
  https://doc.qt.io/qt-5/qtwebengine-overview.html
- changed to latest react-i18next module, as older versions have
  issues with TypeScript, but in latest version experimental
  at-decorators are not supported anymore, so about 25 (mostly
  BitBox01 related) components have to be refactored
- removed Locize in-app editor as not supported with latest version
- started using CSS Module filenaming convention (*.module.css)
- changed Qt build script to search deeper because React build
  nests assets into build/static/js
- removed TypeScript `readonly` instruction as React marks state
  and props as readonly internally
- started removing Preact syntax sugar api

Currently working:
- LanguageSwitch and one translated string as POC
- `make webdev` (with `npm start`) on PORT 8080, custom npm script
  for windows.
- `make qt-osx` desktop app with POC

TODO:
- get tests working
- typed CSS is currently not supported
- migrate from preact-router to react-router

This commit is part of migrating to React, not working on its own.