eth: sync account when making a tx proposal
newTx() accesses account.balance to check if there are sufficient
funds, but that only works if the account is already synchronized.

The eth account unit test failed randomly due to this, as somtimes,
the tx proposal was made before the accoutn synced.

There is a separate issue to be addressed in a separate PR: the
account vars such as `balance` are written and read without a lock
from different goroutines.