frontend: use feerate or gas price for custom fee label
Currently when selecting a custom fee, the label on top of the input
just says 'Network fee', but we use this label also for the total
amount of the estimated fee price below the selection.

Changed the custom fee label to display either 'Fee rate' or
'Gas price' depending on the coinCode of the selected account.