frontend: refactor mnemonicpassphrase component
Added setMnemonicPassphraseEnabled typed api.

Refactored that the component does not depend on the getInfo
callback anymore. This was a callback called after success, so
that the parent component could refetch data and pass it again
via props. Mnemonicpassphrase component is now almost standalone
this way. If needed the props to set the state default could be
removed or made optional, but then the component would need to
fetch default data on mount.