frontend: test for success in api and just return deviceinfo
Some api responses contain a success field that indicate if there
was a problem and would return error message and code in case of
failure.

Changed to check for success inside the api function and return
the success data directly, or reject the Promise and provide the
error code and message.

With this change the success doesn't have to be checked in the
api consumer (component) anymore, and has nicer syntax:

apiSomething()
  .then(() => console.log('success!'))
  .catch(console.error);