frontend: turn caret pointing to bb02 device into a combined icon
Sometimes the BitBoxApp requires the user to focus on the device
and confirm or reject something on the BitBox02.

Recently a combination of an animated caret pointing to the device
has been added.

Changed to combine the icon and the graphic into 1 component so it
can be easily shared and reused. Also used it now in passphrase
on the proceed step, in which the user has to confirm enabling or
disabling on the device.