frontend: add support for multilinemarkup
Similar to SimpleMarkup, this component renders strong elements
in the text into real JSX elements, but with newline support.
Each newline is mapped to return a SimpleMarkup component.

In Preact 8 there is no support for fragments so we cannot
use multilineMarkup as real component, but after migrating
to React multilineMarkup can be used as real component.