frontend: show passphrase configuration in fullscreen
When enabling or disabling the passphrase setting users should
fully focus on the information shown and not be distracted by other
elements on the page, therefore the view in fullscreen is more
suitable than just a dialog.

Move call to replug-the-device to be the last message of the
success view.

Introduced a new <Main> component for contents of the page, this
component should replace <div class="container">. Also left example
code of how <Header> and <Guide> components can be used.

Made also sure the <View> content is scrollable inside <Main>.

Sometimes the BitBoxApp requires the user to focus on the device
and confirm or reject something on the BitBox02.

Turned caret pointing to bb02 device into a shared component, so
it can be re-used in different places like AOPP.