frontend: show passphrase configuration in fullscreen
When enabling or disabling the passphrase setting users should
fully focus on the information shown and not be distracted by other
elements on the page, therefore the view in fullscreen is more
suitable than just a dialog.

Introduced a new <Main> component for contents of the page, this
component should replace <div class="container">. Also left example
code of how <Header> and <Guide> components can be used.

Made also sure the <View> content is scrollable inside <Main>.