docs: add contribution guidelines
As the contributors list keeps growing, we really need to get on
same page in how to propose and make changes to the codebase.
The guidelines in this commit align with what people with merge
permissions to this repo consider common practices.

There are other git repos, such as bitbox02-firmware, api-go and js.
We'll need the same guidelines in those, too. An alternative was to
publish a single copy on a website but having independent
self-sufficient repos was a prevailing benefit. So, will copy these
guidelines over to the other repos once this is committed.

Closes #96

[skip ci]