frontend: passphrase do not forward enabled prop
Currently the device info enabled information is passed from
settings trhough route parameter. The reason was that the info
is already available in settings.

Passing the state with a param is not ideal as the component does
not work on its own, for example when we want to link to it from
the setup success view.