frontend: wait-dialog should allow any children
The wait-dialog component has an interesting logic that checks if
there are no children in which case it displays a graphic that
shows how to confirm on a BitBox01.

The component should also allow just strings as children and then
display that without the BitBox01 graphic. Changed to just check
for children and removed the is valid react component filter.