frontend: fix open exported csv file
The exported CSV file did not open anymore but instead it
changed the view and routed to the wait screen.

This is a regression in Anchor.tsx introduced in:
https://github.com/digitalbitbox/bitbox-wallet-app/commit/c595e1d8a741bb75af190e54d0ae692854c91fe2

Fixed by bypassing Anchor component logic and directly call
apiPost('open', csvfile). But kept the anchor component that the
layout does not break.

But the layout of the open button on the account summary view
looked broken, so I added minimal styles that font and color
matches the rest.