frontend: fix settingsbutton linking
In the Qt app we shouldn't use normal anchors with href attributes
to link to different views. Reason this works in webdev, is that
react-scripts always serves the index.html with main JavaScript,
so after navigation and a full page reload occured the script can
pick up and continue. In the Qt app the index.html isn't served
and therefore something like a 404 happens.

Fixed by chaning all remaining href on settings buttons to onClick
using the deprecated route method. This should be rewritten as
function components with useNavigation in a later commit.