frontend: use view component for bitbox02 setup
The BitBox02 setup styling is a bit out of date, with simple
stupid message signing and enable passphrase there came a new
View component that looks better with bigger font size on larger
screens and flatter without unnecessary white boxes.

Unlock is already using the View component, but the of the bitbox
setup still used an outdated step component. The step component
once showed actual steps, but this changed and steps component
was now only used as a container.

There are a lot of custom styles that are only used in bitbox02
setup and by using the View component the styling can be reused.
View module only need one change to support overwriting text align
to be left aligned in the content area of the view. This is needed
for the success and user confirmation list with checkboxes, that
the layout of those lists do not break.

The step where the user selects between create-new or restore
backup has 2 columns with both options. This could not been done
only with the View component only.

Changed:
- changed all steps from bitbox02.tsx to use View
- added new Grid and Column component to support content that is
  displayed side-by-side in different columns
- use inline (in content) Status and Message component instead
  of Toast component
- attestation failed warning is now shown in the content area
  of unlock and pairing view
- moved the warning graphic to Icons component and deprecated
  a similar icon that is used only for BitBox01
- removed outdated steps component and unused CSS
- moved styles from steps.module.css to bitbox02.module.css
- updated styling of backups component, changed white background
  and padding
- changed secure location icon to use info icon