backend: open CSV automatically after export
SystemOpen() has a whitelist, but the whitelist does not cover
exported CSV files - only the Downloads folder, but the user can save
it somewhere else.

Instead of adding CSV export paths dynamically to the whitelist, it is
simpler to automatically open the CSV. This also avoids the clumsy
hack of converting the button from 'export' to 'open'.