frontend: view component: add option to render as dialog
Last commit changed the alert to always dislpay as a fullscreen
View, but the old components using alert without a view should
keep using a dialog for now unless they are refactored to use
the View component.

Added prop to View to render as dialog and changed alert default
asDialog to be able to opt out.

Marked alert component as deprecated, ideally errors/warnings are
shown in the content area and not in a dialog.