frontend: remove custom fontsize used in status and electrum component
Subheader-mobile was a custom fontsize that rendered very big on
small screen and was only used in the status and on the electrum
view.

Using the big fontsize on small screen was somewhat working for
status boxes that appeared on top of the app, but looked really
strange if inside the view.

Removed the font size, changed electrum settings to just use normal
subtitle, status already uses subtitle font size for medium and
large.