backend/bitbox02: only add script configs that are present in the tx
When signing, the account script configs are provided upfront to the
device, and referenced by the inputs. Before we would add all possible
account configs, even if not all were needed. Now we only add the
script configs that are present in the tx.

The reason is that the BitBox02 firmware decides if the transaction is
a taproot-only transaction by looking at the account configs. If it
is, streaming previous transactions is not needed.