backend/accounts: enable taproot receive addresses for new accounts
Taproot is added to newly made accounts, e.g. the first time a
keystore is inserted or a new account is added.

In a future PR, existing accounts will be updated with taproot as
well.

The order of the script types in the account matters. The first entry
is used at the moment as a default receive type for AOPP, Moonpay and
as the default change address type.