frontend: use dialog to select receive address script type
Toggling/cycling through 3 receive address script types is
confusing now that Taproot is supported, as it is unclear which
script type is currently selected and what the choices are.

Moving to a list of input radios gives a better overview. Using
radio inputs was anyway planed in an upcoming receive revamp.

Also changed radio component that name attribute does not default
to id and can be set by passing throught props, as the name
attribute is used to group multiple inputs.
Radio inputs are only used in the backup component.