qt: add BITBOXAPP_RENDER environment var to configure rendering
We switched to forced software rendering in
daaf2319afb3f5813a4bddd67e13b300590414e5 due to some users suffering
rendering artefacts and crashes with the default rendering backend.

The downside is that the app can be very slow on some machines,
especially when maximized with high resolution.

This env variable is a workaround to enable users to improve their
experience and try the default rendering, which is normally hardware
accelerated.