frontend: show errors from qr scanner
If the app does not have permissions to use the camera the library
returns 'Could not start video source' error.

This change shows the library error to the user in cases something
went wrong, but also if the scaning is aborted and no code has
been detected in that case the library returns 'Video stream has
ended before any code could be detected'.