frontend: improve visual loading of the account summary view
Showing a placeholder (Skeleton) during loading improves the
user experience as there is less layout shifting and clear
indicators where content can be expected.

The skeleton component adds an invisible character so that the
CSS automatically inherits font-size from the parent in most
cases this sets the height correctly, if not there is a fontSize
prop to set the height. The component also has a minWidth prop
to indicate the desired width.

Changes:
- use skeleton for account summary view
- change chart on summary to always take a fixed height
- disable chart buttons until chart is loaded
- render sidebar asap even if subscribed keystores HOC is still
  loading
- just pass data to chart component and let it handle it
- changed from setting accountsPerCoin once onmount to creating the
  map on the fly
- conditionally show buy button only in empty charts

Test:
- rm cache, cache/exchangerates, cache/headers-* or cache/account-*
- start the app and observe loading
- alternative generate a performance report in webdev with the
  lighthouse tool (devtools, select desktop) to see a recording
  of the whole loading from first-paint to fully loaded app

If the wallet is empty it shows a buy button, but if the wallet
ever had data showing a buy button on the summary next to the
chart clutters the layout. Removing the buy on the summary when
there is a chart looks more clean. The buy button is still visible
on the empty accounts.