frontend: add fitContent option to the view component
If a View has a success icon by `<ViewContent withIcon="success"`
but not a lot of other content sometimes the ViewButtons are not
in the visible area on smaller screens.

Refactored the unused `top` option into `fitContent` to indicate
that the View should try to fit all content into the visible area.
With `fitContent` the big success icon will shrink if there is not
enough available space.

Note the footer could still be outside of the visible area and only
visible when scrolling down.

Also added some docstrings to the View component.