backend/eth: add ERC20Balance() to the interface
Before this commit, the balance of the ETH account was fetched via an
Etherscan API call directly, while ERC20 token balances were fetched
using a generated IERC20 package which needs a RPC client to perform
RPC calls to figure out the balance. We provided the same Etherscan
client, which also implements the RPC interface by using the
Etherscan->GETH RPC proxy api calls.

This is a first step in cleaning up the mess, to pave the way to plug
in a different backend instead of Etherscan, such as Blockbook.

Since we don't call NewIERC20() anymore (where the client must
implement `bind.ContractBackend`), we can stop using
`bind.ContractBackend` in our interface, allowing us to remove a bunch
of dead methods. We copy retain the inteface functions that we still
need.