frontend: kerning of progress numbers using tabular-nums
When showing rapidly changing numbers like the number of blocks scanned, tabular numbers make the UI more stable. For example, a number like "11" is narrow while "12" is wider. By using tabular numbers, the width of the UI stays stable when it goes from rendering "11" to "12".

I added `font-variant: tabular-nums` to a few progress indicators. There is one indicator in account.tsx (`initializingSpinnerText`) that I didn't modify since it would require bigger changes to the Spinner component, such as allowing a parent to pass in JSX children instead of supporting only the `text` prop.