frontend: improved insufficient gas error for erc20 txs
If the user tries to send an erc20 transaction without having enough
ETH to pay for the fee, etherscan returns a rough error message, which was
previously displayed in the frontend as is.
Now the backend returns a dedicated error code and the frontend displays
a comprehensible and localized error message.