frontend: modify fiat historical value placeholder in tx
Fiat historical value is shown in tx detail. If the transaction is
recent the value could be not available. The placeholder we used was
'---' which led to poorly readable results for incoming/outgoing txs
('+---' or '----'). Now, if the amount is not available the placeholder
is shown without the sign, which should be more readable.

In the future it would be nice to refactor the FiatConversion component
and limit the `sign` prop to '+' | '-'.