frontend: refactor fiat default
The current styling of the default fiat is star becomes orange
(outlined icon) and the toggle becomes inactive as it is disabled
so that the default cannot be turned off. The styling looks
confusing as it is not clear that this is the default.

To improve the visual styling this commit:

- added small '(default)' text after the currency
- made active star more prominent, changed from outline to filled