frontend: add translation feedback link
As we are continuously adding and updating translations, we should
have a way to give feedback for the translations to help us catch
translation errors.

Added a mailto link at the bottom of the guide and allow mailto
links to be opened by the backend.

Keeping the translation feedback english-only for now, as it is
better if we get feedback in english than in the native language.