frontend: fix coin and unit styling in tx overview
Sometimes the unit breaks onto a separate line, this happens with
large amounts or coins with lots of decimal for example with ETH.

Changed styling so that very large amounts are cut off with "…",
but only visually so that copy/paste contains takes the whole
amount. This should also fix the issue with large amounts on
small screen, in extreme cases the fiat amount should squeeze
the coin amount together.

Also added "font-variant: tabular-nums" definition for the amount
so that all numbers take equal space, without using a monospace
font.