android: add external link queries for Android 11 permissions
With Android 11 the visibility of the installed packages has been
filtered to offer a higher privacy level to the user. This commit add
to AndroidManifest.xml the `queries` elements needed to open external
links for web pages and `mailto` urls.

Queries support is not granted in all Gradle versions. For this reason,
the latter has been updated to v3.5.4.