frontend/electrum: split ElectrumServer into two components
It was used in two modes, adding a server and showing a server. This
means there are states and props used in in one mode and not the
other, which is confusing.

Splitting into two components that only do one task each is clearer
and easier to convert to TypeScript functional components.