frontend: converted LanguageSwitch to a function component
As a part of our refactoring efforts and to follow the latest
conventions in the frontend, class components need to be
converted into functional ones whenever possible. Tests are
updated so they work with the newly created functional
component and its hooks.

Also migrated the test away from enzyme to testing-library/react,
as we plan to remove and not use enzyme anymore in the future.