frontend: fixed download cert, check, and add buttons
the ‘download cert’ and ‘check’ buttons were not showing the
loading animation when they were  supposed to. The problem
was due to the buttons' incorrect styling. It caused the
animation to look invisible.

to fix the issue, their style has been altered to be the
same as the buttons at the top of the page (the ‘Check’ and
‘Remove’ buttons) The `disabled` conditions for each button
have also  been slightly refactored to increase its robustness
& readability which also fixed an existing bug.