frontends: color variables refactoring
Separated text and background colors into their own variables. This
makes colors easier to maintain because now it's possible to change
text and background colors without affecting each other.

Also some text used --color-softblack instead of a designated
/* font colors */ variable. Changed this text to --color-default so
it is consistent with the other text colors.

- changed text using color-grey to color-secondary
- created --color-tertiary and --background-tertiary that use
  --color-gray-alt
- removed unused text-darkgray, text-black and text-bold

TODO: for the future --color-blue to --color-info