frontend: changed return types & children types for FC
As a part of our refactoring effort in the FE to increase
our code's readability and consistency as well as following
the best practices and conventions, we'd like to change several
things mentioned below for our functional components (FC):

1. If a component optionally accepts children we keep using
`FunctionComponent` as we do already (or start to use one
when it isn't used yet).

2. If a component does not accept any children we remove
`FunctionComponent` and `PropsWithChildren`.

3. Remove return type `JSX.Element` or `JSX.Element | null.`

4. Children's type is mostly `ReactNode`. When component must
only have exactly 1 child, the type is `ReactChild`.