backend/eth/account: remove account lock, add intialized lock
Locking the whole account is confusing, as it is not which struct
fields the lock covers and when to use it. Individual field locks are
much easier to reason about.

CanVerifyAddresses does not need the init check, it works regardless.

GetUnusedReceiveAddresses needs the init check as account.address is accessed.