frontend: use view component for device settings
Settings can use View, Grid and Column component for layout instead
of plain div's. The View component is mostly used in fullscreen
mode but an early version did support normal mode with sidebar and
guide. Fixed some minor CSS issues (spacing etc.) so that normal
mode should now be fully working.

A differnece is the position of the title element, this is now in
the content using ViewHeader component.