frontend/pocket: add address verification
A safe buy workflow requires a verification of the address that has been
shared with the exchange. In this case pocket widget sends an email
containing the address that has been shared by the user and requires
a verification via `postMessage`. The address is displayed on the bb02
and should be ideally verified by the user opening the email con a second
device (e.g. the phone) for better security.