frontend: fix view component with guide bug on large screen
On large screen the guide on device settings broke onto a 2nd line,
this is a regression introduced in:
- 484c42f59207f12f428f588935cd3628e603d83b

Fixed by adding the old solution with 2 containers but as react
components. This does duplicate some CSS, but in the long run,
using the global CSS should be phased out.

Added overflow auto to the '.container' CSS, so that longer content
is scrollable. This should only impact the new react components
that is used in device settings and not have any side effect on
other views.