frontend/pocket: remove address sign abort alert and refactor
Previously, when user aborted address sign in pocket workflow, a popup
alert was displayed. Now it is removed, since it is not necessary and
pretty ugly :) There is already a clickable text at the bottom of the
widget that allows the user to try again.

This commit also fixes a bug that allowed to user to send more than one
address signing or verification request at the same time.

The error handling is also refactored following the same pattern done in
the address verification.