frontend: created "Receive" btn for empty wallet CTA
We'd like to have a "receive" btn (instead of only "buy")
in the CTA as it's a valid way of depositing coins.

1. Modified the buyCTA to also able to include "Receive" btn.
This is displayed in the account summary screen.

2. Extracted the account selector into a single component so we
can reuse it in the new receive screen and still use it in the
buy screen.

3. Created a new receive screen/page (`/select-receive`) which
displays the said selector.

4. If there's only one account, upon clicking the "Receive" btn
user won't be broguht to `/receive`, but instead directly to
their only account's receive screen.

5. on empty account(s), layout changed. Hiding table and
default buttons to send, receive and buy, and show the
new buy & receive CTA under the balance.