frontend: fix device settings flicker
The whole app window flickered when the user navigated to the
device settings, including the sidebar. Reason: device settings
route first load: deviceswitch that checks if device is bb01, bb02
or bootloader, for bb02 next the bitbox02.tsx is loaded, that
also handles unlock, pairing, create and restore, this component
does many api requests and is most likely the reason for the
flicker as it takes some time and will eventually load the bb02
settings component.

Involving deviceswtich/bitbox02 before loading device settings
seems to be the cause for completely blank view if for example
btc nodes are misconfigured, probably due to accounts not being
initialized correctly.

This change adds a new settingsswitch component that can load the
bb01 or bb02 settings component directly without going trough
deviceswitch.

Drawback: some shared components such as manage-backup or sdcard
check have a back button, this should not leave the setup if the
component is used in setup. therefore keeping the back buttons
pointing to deviceswitch.