frontend: fix regression on Dialog and created dialog-legacy
During the development of the new Dialog (improving dialog
for mobile), we introduced a regression that affected
several places such as the:

1. Factory reset device
2. send-QR

This PR fixes it.

We also will use the old dialog for all of the bb01 components
and the Confirm component (which uses dialog internally).

This PR also fixes some other random bugs that are not exactly
in the Dialog component but still related to it, such as:
1. WaitDialog style on mobile
2. Closing send-QR with ESC only toggles the dialog, not
actually closing it.