frontend: use native-locale to pre-select region dropdown in exchange page.
We have preivously used native-locale but we didn't take
into account that native-locale may sometimes return
nothing (empty).

Another PR then was created to fix this, using
`i18n.language` - which is not exactly correct.

Thus, in this PR we used native-locale back. But, we are now
handling empty native-locale gracefully so it won't crash the
page whenever native-locale returns empty/nothing.