backend/pocket: refactor `PocketUrl` to use devservers argument
This updates`PocketUrl`to return the address of the Pocket widget from
staging or production environment depending on the `devservers` flag,
instead of the net used, which was the previous approach.

It also adds a `servewallet-mainnet-prodservers` target in the Makefile,
which runs the server on mainnet and with devservers flag disabled.