backend/usb: run all hidapi API calls in a dedicated OS thread
For unknown reasons, on macOS, there are intermittent errors during
hid.Device.Write():

`IOHIDLibUserClient:0x100002065 setReport failed: 0xe00002bc` In the
BitBoxApp, the issue appears most frequently in the loop polling the
device for a API call status in query.go/rawQueryV7.

Running all calls in the same OS thread seems to resolve the issue,
even though it is not clear why that is. Maybe something in the macOS
SDK relies on thread-local variables, or there is a memory-bug that
does not get triggered this way.

After a long time debugging this, I am still not sure what causes it,
especially because it also happens in a test program without any
goroutines of this sort:

`main() { for { C.hid_write(...); C.hid_read(...) }`.

The issue disappears however if there is a `runtime.Gosched()` in
every loop iteration in this test program, which is perplexing, as
there are no other goroutines to yield to. Further investigation is
needed to figure out what is going on under the hood - for now we
apply the fix to run all calls in the same thread. `runtime.Gosched()`
is not needed in our case as the scheduler is automatically invoked
when sending to and reading from the `funcCalls` channel.

As an side: the cgo code generator does make sure that the goroutine
calling a C function (hid_write for example) is fixed to an OS thread
during the execution of the function, so it really seems like multiple
such function calls need to run in the same OS thread.