frontend: fix scanning qr code with amount in sats mode
The send interface shows an error when scanning a QR code with amount
when in sats mode. Reason: sats cannot have decimals on L1.

Added a check to see if amount is specified and app in sats mode,
in which case it tires to parse the amount and multiply with 100m.