android: run go backend on a dedicated service
Before this, the Go backend was invoked on Android throught the
`GoViewModel` class. We recently discovered that this solution could
cause timeout errors if the app loose the focus during a polling to the
bb02, e.g. while the user is verifying an address.

The Go backend is now started and executed inside a dedicated
`GoService`, bound to the `MainActivity`.
When the service starts it immediately promotes itself to foreground
process, to mantain a good priority even when the app is out of focus.
In order to go foreground the process needs a dedicated notification,
that is set to a low priority level to be as less intrusive as possible
(basically a text in the taskbar that says that the app is running).