frontend: fix "back btn" behaviour on country selector (buy page).
Back button in the country selector (`buy/exchange/`) page
doesn't work as expected when there's only one account. This is
due to the back button is trying to go back to `/buy/info`, but
that page redirects the user back to the `/buy/exchange/` page
again. Therefore no "back" action is really happeing for
the user.

We fix this by going back 2 pages when there's only
one account, and only going back once when there's more than
one account.

`useNavigate()` is utilised for this approach.