frontend: set dark mode colour scheme for account dropdown/selector
Hardcoded CSS values to display the account dropdown in dark mode.
We added some selectors in the css file, where previously weren't
needed.

This is because most of react-select's default styles
suit our light mode. However, it's not the case for dark mode.
Thus, we need to grab ahold of more selectors to be customised
(hence adding more selectors in the css).