btc: enable custom fees w/ info text if fee targets are missing
In case the fee estimations were missing for any reason (e.g. download
too slow or some error), the send screen was unusable, as the
transaction proposal could not be made due to missing fee. The error
was not propagated to the user, and a transaction could not be made.

This patch enables custom fees as a fallback in that case, even if
custom fees are otherwise disabled in the settings. It is not
great (probably confusing for most users), but better than not being
able to transact at all. Luckily this should not happen often - the
only instance we know of so far has been with a user that used a
custom Electrs node on a slow connection, going into the send screen
immediately after sync.

Alternatives considered:
- try downloading the fee rates on the fly if they are missing. Didn't
  do this as we currently don't have custom Electrum request
  timeouts. In case the fee rates were missing due to the server or
  connection being slow, this would block loading the send
  screen. Doing this approach in a good way would take much more time,
  and since it is a rare problem, I opted for the quick fix.

Co-authored-by: thisconnect <this@shiftcrypto.ch>