frontend: change some assets via css classes
Some assets are used in the settings view and did not change on the
fly, reason is that darkmode is currently a global variable and
does not trigger a re-render.

Proper way would be to use either react context or some global
app state library such as mobx or redux.

In the meantime this commit uses a simple show-in-darkmode and
show-in-lightmode class to hide/show the correct element. Drawback
is that there are always 2 elements (one for each version) in the
DOM.

Used this workaround on assets that are used on settings, which are:
- dialog close button
- hamburger
- swiss opensource

also fixed SwissMadeOpenSource component that it properly added
the className to the image.