unit-tests: avoid hitting real Electrum servers and Etherscan
In aopp_test.go, a full backend is used, and the accounts created use
real (unmocked) coin instances connecting to real production Electrum
servers / Etherscan.

This results in fragile unit tests, as they can fail if the network is
down etc. The unit tests then would hang forever, waiting for
synchronization to finish.

We mock the relevant blockchain connection instances to avoid this.