backend: sort accounts in Accounts()
We have persisted accounts in accounts.json and load them into live
accounts.

We want the accounts sorted by coin, account number for the
frontend (sidebar, manage accounts). Until now we sorted the accounts
while loading them from accounts.json.

It is preferable to sort the accounts only when serving them to the
frontend - then we could easily modify the backend.accounts list on
the fly and add accounts without having to call `Reinitialize()`,
which closes all accounts and reloads all of them, just so that the
sorting is correct. When we add automatic accounts discovery, we don't
want to Reinitialize() repeatedly for every account that is found, but
just add it to the list.

The sorting function is now a bit more complicated because it needs to
also sort ETH with their ERC20 tokens explicitly (before, they were
correctly ordered implicitly by the order in which they were loaded
from accounts.json).