backend/environment: detect OS theme at startup
On Android, the preferred theme is handled using media queries in the
frontend. This allows to update darkmode setting on the fly if the user
enables a dark theme at OS level.

On Qt this doesn't seem to work correctly, as the media queries return
weird values. This update creates a dedicated `detect-dark-theme`
endpoint that exploits backend environment functions to apply the
correct theme at startup under Windows, Mac OS and Linux (gtk, kde,
xfce4). Note that this doesn't allow on the fly updates of the theme.