api/bitbox02: allow backups using recovery words, skipping sdcard
This adds a parameter to the createBackup endpoint to choose the
sdcard backup or the recovery words backup. This can be then used
later to allow skipping the sdcard backup in favor of the recovery
words backup in the BitBox02 setup wizard.

Since this feature is only possible since firmware v9.13.0, one must
first check if `versionInfo.canBackupWithRecoveryWords` is true before
using the recovery words method.