frontend: setup with recovery words only
Add support for creating a wallet without a microSD card backup
using only recovery words.

Opt-in to words is currently a toggle under create wallet, but this
is subject to change and might get burried in some expert config.

Also create backup with sdcard has a warning step with a list of
checkpoints. This is currently not shown for creating with words.

Also changed to go back to choice of create and restore from when
user aborts the process.