frontned: created tabs and equivalent mobile menu for the new revamped settings page
On desktop, we'll have tabs for each sub-page of the new settings page.
On mobile (small screens), we'll hide those tabs and have back button on those sub-pages instead.

Additionallty, on mobile, we'll have a general / index settings page which will have the
settings menu. This page will not be visitable for users on desktop.