backend: simplify RenameAccount
The `account.Config().Config` is a pointer to the persisted account
stored in `backend.config.accountsConfig.Accounts`.

By making the the slice entry also a pointer, we can make sure that
the `account.Config().Config` always points to the persisted
config. Then, we can simply modify the persisted config to also modify
the loaded config.

Before that, the same was true but not deterministically: the stored
account config and the loaded config would point to the same memory
location, but only as long as the persisted account was not moved,
which can happen when the slice is extended (so the slice and all
contents is moved to a new location while growing the slice). If this
happens, then the loaded config would become a shallow copy, which is
error prone: modifying a primitive field would modify a copy, while
modifying a pointer field (e.g. the signing configuration) would
modify both the loaded and the persisted config.

The alternative to this commit would have been to instead make a deep
copy of the persisted config when loading it. It seemed to me like just
sharing the same memory here is simpler overall.