frontends: refactor accountssummary subscriptions
Accountsummary may not have all accounts on mount and needs to
re-subscribe on componentDidUpdate.

Simplifying the code a bit by removing map, return and a Promise.all
that did nothing, seems to be a leftover.

To test this delete account cache before servewallet:

    rm -rf appfolder.dev/cache/account-* && make servewallet

Next step is changing apiWebsocket to syncdone and statusChanged.