frontend: move bitbox02 restore-success views to own module
Moved the 3 success views to success.tsx. Only the create-wallet
success has the nice success animation. The other 2 success views
show a list with security notices. This could be improved in the
future so each has the nice animated checkmark and different
success messages.

This is the first commit moving each setup step to its own
component so that it is easier in the future to make changes to
the setup.