frontend: move channel-hash-changed subscription to pairing
The channelHashChanged subscription, hash and deviceVerified state
do not need to be in the parent component. Moved to pairing
component.

This could be further simplified if the subscription would contain
the actual data (hash and deviceVerified) directly, but currently
subscriptions are only used to notify the frontend that there is
an update, so the frontend has to fetch this data again.

Also changed order of attestation failed warning so it is the same
(after header) as in password entry.