backend: fix ETH balance update including pending txs
ETH accounts update fetches the balance from etherscan. This amount does
not consider possible pending outgoing transactions and this means that
after sending a txs the balance is not correctly updated, until the txs
gets confirmed.

With this, the balance fetched from etherscan is decresed considering
the amounts and fees of pending outgoing transactions.

NOTES: The fees are only considered when computing the balance for ETH
accounts, since for ERC20-tokens they should be detracted from the related
ETH account. This is not implemented yet, as it would require a link
between the ETH main account and the ERC20 derived account. It could be
possibly solved with a future migration to blockbook chain explorer,
that could allow to list the ERC20 pending transactions within the main
ETH account transactions fetched.