frontend: remove unnecessary app state change
Seems like setting appState to createWallet is not needed anymore,
was introduced in fcce3e87157bd0e9335ab75a8420bab16e9016e2

This subscription is fired after successful bitbox02.setPassword
but then app state is already in create-wallet. An edge case
would be if the user closes the app after successful setPassword
and reopens it, but this should actually go back to setup choices.