frontend: added guides for relevant settings pages & modified header
There were several changes to make the guide works with our
tabbed settings.

1. We now use the `Header` component's prop `title` for showing both
setting page name for both desktop & mobile.

e.g "Settings" on Desktop and "Appearance" on mobile
(when on `/new-settings/appearance`). Previously, to display
the page name on mobile we use `WithSettingsTabs`

2. Modified `WithSettingsTab` to not receive the page title
anymore due to the point above.

3. Modified `Header` to be able to hide sidebarToggler also due
to point 1.

We need to be able to hide `sidebarToggler` because
on an individual setting page on mobile (for example `Appearance`),
we already have a back button which is inside `MobileHeader`.

So we don't need the `sidebarToggler` anymore.