backend: update to v4.37.1 (hotfix)
Some users reported the app crashing when Litecoin is enabled with the
the following error:

> panic: readScript: script witness item is larger than the max allowed size [count 65973, max 11000]

The limit in question was fixed by btcd a while ago, so we simply
update btcd.

It is unusual to have such large witness sizes, so there is a chance
something else is going wrong that casuses the transaction to not
deserialize correctly.