frontend: block toggle account until coin is enabled/disabled
There can be a latency as currently all accounts are re-initialized.

This change sets the toggle to disabled until the enabling is
complete. To keep it simple this doesn't keep the disabled state
of each toggle, but just fire and forgets.

Setting the toggle to disabled quickly changes the color. In a
future commit this could be changed a loading indicator  instead of
chaning it to disabled color.