frontend: change conditions to show BuyReceiveCTA
showing BuyReceiveCTA properly to avoid blank screen on account
screen on testnet.

Now, BuyReceiveCTA on the account screen will show if account is
empty. Only when the exchange supports the buy of that certain
coin, the BuyReceiveCTA will render the Buy button along with
a Receive button. When it's not supported, only Receive button
will be rendered. The same will happen in account summary.