backend/bitsurance: add bitsurance backend logic
This is the first step of the integraction of a BTC account insurance
feature provided by our partner Bitsurance.

This commit includes the backend logic to:
- compute the xpub salted hash used as an anonymized accout Id in the
bitsurance accounts ledger.
- fetch the most up-to-date list of insured accounts hash and monitor
the insurance status of the enabled accounts.

The service is integrated keeping in mind the privacy of the users,
minimizing the calls made towards bitsurance servers (i.e. no call is done
until the user manually request a check of the insurance status of the
accounts).