frontend: refactor firmware-upgrade-required to functional component
Changed FirmwareUpgradeRequired to use View component, also with
this BitBox02.tsx does not need to use translation decorator
anymore and types can be simplified.

Another change is that the whole view is now in fullscreen, as
sidebar/settings is not really needed here.

Test by increasing lowestSupportedFirmwareVersion in both files:
- bitbox02-api-go/api/firmware/device.go
- bitbox02bootloader/firmware.go