frontend/bitsurance: first integration of the service
As a follow-up to the previous commit, which included the first part of
the backend logic needed to integrate the Bitsurance service, this adds
the frontend handling of the same logic.

- A new page bitsurance.ts is added, that allows to manually check the
insurance status of the active accounts.
- An automatic check (only on already insured accounts) is also added,
that shows an 'insured account' badge on the account page and alert the
user in case of modifications of the status of his policy (e.g. expiring).