backend/lightning: add the lightning handler
This handler creates the API endpoints to handle interaction with the Breez SDK/Greenlight lightning node. When the account is initialised with Init, the handler's Init should be called to initialise the associated Breez SDK instance, which internally calls connect. Any events received from the Breez SDK instance are passed to the frontend via subscription callback