backend/bitsurance: switch insurance status endpoint
Previously we fetched a list of all the insured accounts using the
`allAccountHashes` endpoint. This solution is not much scalable and forces
Bitsurance to expose a list of all the active accounts. This commit
exploits the `accountDetails` endpoint instead, and makes the app do a
single endpoint call for each account to be verified.

This also adds a test for `BitsuranceGetId`.