frontend: use qr-scanner
Changing library to scan QRCodes from @xzing/library to qr-scanner
which is a more modern port of the same library, see
- https://github.com/nimiq/qr-scanner#readme

The original @xzing/library project is in maintenance mode, with
no active development, see:
- https://github.com/zxing-js/library

Another problem was that at least our version of @xzing/library
could not scan modern WC V2 Qr Codes.

The new library also has a highlightScanRegion feature to render
an outline around the scan region on the video stream.

Unfortunatelly the dynamic import did not the same way as with
the old library, but it seems this only slightly increases the
build size (see below). The issue was a combination of lazy
loading the library, still depending on types from it and passing
a react DOM ref, which made everything complicated.

Dynamic loading could probably work on a higher parent component,
i.e. the whole receive-address or even the full view Send route.

Adding the QR loading spinner "as is" without conditional
rendering, as it seems to unmount the video component when a state
changes in the component (scan-qr-video).
As there's no loading state, the spinner appears initially before
the QR reader starts with the video. Once the video loads,  the
spinner disappears on its own.

make buildweb before (master 897dd97d2)

  644.14 kB  build/static/js/main.06fd36e9.js
  87.22 kB   build/static/js/459.4bf01830.chunk.js
  35.56 kB   build/static/css/main.9773ffd0.css
  157 B      build/static/js/909.925590b9.chunk.js

make buildweb with this commit

  649.18 kB  build/static/js/main.d274d59b.js
  35.59 kB   build/static/css/main.65886e68.css
  10.32 kB   build/static/js/684.891a2f0a.chunk.js

Other changes:
- increased the size of the QrScanner video popup on desktop

Co-authored-by: shonsirsha <seangeekpro@gmail.com>