frontend: use context to initialise & share wc's web3wallet obj
in order to fix & prevent bugs coming from web3wallet
by walletconnect, and adhering to react's best practices, we
should move away from exposing it as a global variable.

Instead, we should use try to use React's context API which is
designed for this.

Also removed WalletConnect PoC files and other unused files
in this commit.