frontend: group accounts by keystore in sidebar and account settings
Group shows the keystore name.

The BitBox02 keystore name for now is simply the device name. For
passphrase users that means that multiple keystores will have the same
name. We will not give this any special consideration for now. The
user can label the individual accounts, which is enough to
disambiguate.