backend/account: add ConnectKeystore() api call
This waits for a the account's keystore to be connected. Useful in
watch-only when an action requires the keystore, like displaying an
address or signing a transaction.

The backend will invoke this wherever the keystore is needed.

A frontend call to invoke this is added as well for the cases where
one wants to prompt the user to connect the keystore before the
backend actually needs it, like:

- when verifying a receive address, the user should be prompted before
  the dialog with the full address and QR code appears
- when clicking 'Review' in the send screen, the user should be
  prompted before the review dialog appears

Maybe there are more instances like this, which will be fixed in
future commits.

The minimum Go versio is bumped to 1.20 as `context.WithCancelFunc()`
was introduced there.