frontend: optimise wc loading and web3wallet initialisation
every time a user visits /wallet-connect/dashboard, the app
initialises the web3wallet object, achieved through the use of
Core and Web3Wallet, which are lazy-loaded. The function is
initializeWeb3Wallet.

Only on a successful dapp via WC connection,
hasUsedWalletConnect is set to config. This condition is
essentially used to check if we should call initializeWeb3Wallet
on app start (via provider), or on dashboard first visit.