frontend/bitsurance: first bitsurance workflow integration
This creates the Insurance menu item, that leads to the bitsurance workflow
starting page. From there, it is possible to detect already insured accounts
or begin the onramp of a new account insurance. A bitsurance Iframe is
integrated in bitsurance/widget.tsx, with a structure which is very similar
to the Pocket one.