frontend: fix camera on linux
Regression introduced in 48eebe98e81fa1c026c45e3b9ed1b1f602aae435

That commit added a please-grant-camera-permission warning, in the
same component that contains the video element. It uses a new hook
to check for camera devices and conditionally renders the warning.

Conditionally showing the warning based on the hook re-renders the
component and seems to be the cause for the video to error with:

    DOMException: The play() request was interrupted by a new load request.

This commit is an approach to lift the warning up to the parent
component. It seems to work on macOS, but the warning does not go
away.
Could be that the event listener on permissionstate object is not
working.