frontend: revert "display warning if camera has to access to scan"
This reverts commit 48eebe98e81fa1c026c45e3b9ed1b1f602aae435.

That commit added a please-grant-camera-permission warning, in the
same component that contains the video element. It used a new hook
to check for camera devices and conditionally rendered a warning.

Conditionally showing the warning based on the hook re-renders the
component and seems to be the cause for the video to error with:

    DOMException: The play() request was interrupted by a new load request.

Looks like that was the reason why Qt Linux didn't load the camera
but for some reason it worked fine on macOS.