frontend: locate single back button on the left in ln send
A single back button in ViewButtons should be placed in reverse
order. This is the first time we use a back button only without
a promary action, as the UI automatically moves to the next step
when an invoice is detected.

Updated ViewButtons to accept reverseRow prop, to flip the order
in this case. This has no impoact on small screen as the buttons
are not rendered in a row, but in a column.