frontend/android: block unexpected external requests
Commit 3b62772 introduced a base url to be used in external requests to
comply with Moonpay frame-ancestors content security policy. It somehow
broke the update of the android webview location and made impossible to
block external requests on the base of the page that originated them.

This commit implements a workaround that updates the location after
navigation related events and blocks unexpected requests as already done
in Qt.