backend/keystore: fix notification order on keystore deregister
On keystore deregister the accounts linked to the keystore were closed
and sent statusChanged events to the frontend + a final `accounts` event
to notify the change in the accounts number. The lag between the status
and the accounts events caused the execution of calls towards endpoint
of already closed accounts, which generated unexpected errors.

This send an accounts event immediately after each account close, fixing
the lag.