frontend: fix empty accounts message
After chaning account settings and navigating back to account
summary, sometimes no chart and no empty-wallet message show.

Reason seems to be that in this moment the frontend has all
balances and wrong condition on when to render the message.

Changed and simplified the logic a bit. It should now show the
message unless there is either a balance or one of the accounts
once had coins.