frontend: make qrscanner fullscreen on mobile
- use view dialog component for qrcode detection

Previously it was using the old dialog component, which worked ok
on mobile, but for some reason rendered too big on large screen
with a vertical scrollbar.

- changed to use View component as dialog.

Onchain qr scan video is now fullscreen on mobile. Also Refactored
walletconnect to use view components.

- refactored useQRScanner hook

Send has many state updates and the hook did not properly unmount
this was causing the video to flicker and stop after rerender in
webdev.

Refactored to store the instance of QrScanner in a ref to keep the
same instance when re-rendering.