backend/accounts: do not reload current keystore's accounts
When doing a soft-reload, we do not need to unload and reload the
accounts that belong to the current keystore - we can just leave them
as is to speed things up.

Without this, when disabling watch-only while a keystore is connected,
all accounts would be unloaded and the keystore's accounts loaded
again. With this, the keystore's accounts are not unloaded.