watchonly: add global watchonly setting
With option to exempt individual accounts.

- When global watchonly is enabled, all currently loaded accounts and
accounts loaded in the future will be set to watchonly. Accounts that
had been persisted beforehand are not automatically set to watch-only.
- When global watchonly is disabled, all persisted accounts' watchonly
status is reset. This way, when one enables watchonly again, accounts
that are already persisted don't immediately appear - one has to first
load them with a keystore.