backend: scan accounts when connecting keystore
Before, we only scanned more for more acocunts when accounts were
added, which used to happen when a keystore was connected.

With watchonly, accounts could already be loaded before any keystore
is connected. When a keystore is connected, the accounts are not
reloaded. New accounts would not be scanned. The keystore is needed to
scan accounts as the xpubs of the additional accounts are required..

In addition to scanning accounts when an account was added, we also do
it when a keystore is connected, to invoke scanning also when a
keystore is connected for accounts that already were loaded with
watch-only before.

The unit tests are modified to count only the shown accounts, as the
new call adds some hidden accounts that are not relevant to the test a
bit earlier.