backend/accounts: discover/add first account as well
We have automatic accounts scanning to discover additional accounts
that have a transaction history. However, if no account of a coin
existed, it was not scanned for at all.

This is an edge case - normally one account of each supported coin is
persisted as part of the default set of accounts. This bug only really
appears if e.g. a user switches from BitBox02 BTC-only to Multi, in
which case e.g. the Litecoin accounts would not be automatically
discovered.