build/os-x: fix build errors for M-based Apple machines
Apple machines with M processor have an arm64 architecture with a
compatiblity layer for x86_64 software. Brew packages are installed
with an arm64 architecture by default on these machines.

BitBoxApp is compiled for x86_64 arch, and the arm64 Qt was causing
errors in the build process.

This commit introduces an osx-brew.sh script that forces the installation
of x86_64 deps even on arm64 machines, fixing the error.