backend: fix regression related to updating persisted accounts
It was supposed to only be applied to the accounts that belong to the
current keystore, but 52bf446c7e1d09ac95c7c4d61e9f784fe028e1c1 added a
regression and it was applied to all persisted accounts.

The `setWatch()` thing can be removed - it was there before when the
watchonly setting was global and accounts would need to be marked as
`Watch` when the keystore is connected. After we moved it to be a
keystore setting, it can only be enabled after connecting the keystore
anyway in 'Manage accounts'.