frontend/utils: sort getAccountsByKeystore map
The input accounts list seems to be possibly ordered differently in different
places in the code (e.g. sidebar vs manage account page). This ensures that
the list of accounts grouped by keystore always has a consistent order, and also
keeps keystores with same name (e.g. create using different passphrases) to be
close to each other in the list.