frontend: fix 100vh in mobile webkit
Some elements display larger than 100vh on mobile WebKit,
causing for example the Settings button in the sidebar to be
out of the view.

WebKit should support -webkit-fill-available so that the correct
max height is set, keeping 100vh as fallback.

frontend: test larger base font size

Some elements such as select dropdowns zoom in on focus on
mobile WebKit.

This is only a test.

frontend: fix sidebar height on webkit

Bottom has still a gap.