frontend: bip-85 settings
Added an BIP-85 settings view to explain the most important points
about this feature.

Additional changes:

Relaxed ViewButtons children to be optional, reason: this way we
can keep the component in the DOM even when emtpy so that it does
not do a layout shift.

Changed settings item flexbox styling so that both container can
break onto 2 or more lines (small screen). This happenes when the
settings item shows firmware-update-required.

Added option to center the text of one column.

Allow links to Bitcoin's bip repo.