frontend/keystore: create keystore context to handle keystore requests
With the current implementation, when the wrong keystore is connected after
a keystore request, the app is not able to automatically detect the
connection of the correct keystore. The only option for the user is to
cancel the 'connect keystore' popup and repeat the action to create a new
keystore connection request.  This is counterintuitive and inefficient.

This commit introduces a frontend keystore context that allows
- to keep track of keystore connection requests in a centralized way
- to easily create keystore connection requests in the frontend, providing
a callback function to be called in case of success
- to automatically detect and handle the connection of the correct keystore
after a wrong one

As a first use example, the keystore context is used in the buy/info.tsx page.
All the other occurrences of `connectKeystore` will be replaced with the new
method in the next commits.