frontend/account: use keystore context provider
This removes the `connectKeystore` calls from `ActionButtons` and `Receive`
components and uses the `requestKeystore` method, exposed by the keystore
context, instead. It also adds the 'requestKeystore` in the pocket widget
page, where it was missing.

There is still one invocation of `connectKeystore` that should be removed in
favor of `requestKeystore`, but that will need to convert the `Send` component
into a functional one first.