frontend: sort accounts by rootfingerprint and account code
Sort by keystore.rootFingerprint and account.code, this should
fixes and adds a test where multiple keystores have the same name.

Technically account.code would be enough as it includes the root
fingerprint, but we can't assume that the code will always stay
that way and should not depend on it.