frontend: created and used GroupedAccountSelector
to improve our UX when multiple keystore is connected,
we'd like to add a grouping on the account selector.

This account selector is used on the info buy page (`/buy/info`)
as well as the receive page (`/accounts/select-receive`).

frontend: created and used GroupedAccountSelector

to improve our UX when multiple keystore is connected,
we'd like to add a grouping on the account selector.

This account selector is used on the info buy page (`/buy/info`)
as well as the receive page (`/accounts/select-receive`).