frontend: validate devicename input field in setup
We already have a backend validation, but that only shows an
invalid input message after the user clicked continue.

Better experience is to give live feedback and information about
what is wrong.

It should only allow: printable ascii chars, with max length of 30
including spaces, but not starting or ending with (that is just
trimmed, without showing an error to the user). And show invalid
chars if there is any.