frontend: set device-name before starting restore from mnemonic
Restoring from mnemonic did not allow to set a device name so it
was always 'My BitBox'.

Added set device name step before continuing with restoreMnemonic.

This also fixes a bug in webdev that was caused by calling
restoreMnemonic on mount, ending up with multiple api calls and
a fail message after the workflow was done.