frontend: update dev dependencies
'npm audit fix' reduced high severity vulnerabilities from 3 to 2.
With the other changes at the time of this commit the vulnerabilities
are down to 0.

updated vite from 4.x to 5.x. Tested qt-osx, weblint, webtest and webdev.

Looks like we are not affected by any of the breaking changes:
- https://vitejs.dev/guide/migration.html#migration-from-v4

updated vitest from 0.34 to 1.2.2

Looks as we got lucky and are not affected by breaking changes,
note this version only works with vite >= 5.x

- https://vitest.dev/guide/migration.html

updated jsdom from 22.x to 24.x

jsdom is used in vite tests:
- https://github.com/jsdom/jsdom/blob/main/Changelog.md#2400

various minor versions updated.