frontend: replace BrowserRouter with HashRouter
Using the BrowserRouter is needlessly complicated, we don't need the
browser to make a new "request" when we route to a different
page. Getting this to work is especially difficult in iOS, where the
file path to index.html is loaded in the webview, and routing to
e.g. `/account-summary' fails because the absolute filepath
`/account-summary` does not exist and is outside the app folder
sandbox.

Using the hash router instead, where the app path is tracked behind
the `#` (e.g. index.html#/account-summary) is more appropriate and
easier to handle.