backend: Add lightning package
Add lightning package to support connectivity to Greenlight via the Breez SDK. When activating lightning it uses the keystore's deterministic entropy to create a mnemonic and stores it to the lightning config. Using handlers it maps JSON requests from the frontend to the Breez SDK and returns a JSON map of the response data. Any events received are emitted via observable events to the frontend.