frontend: fix bitsurance button in german
In German the button text was much longer and broke onto 2 lines.

Also the button icon was not positioned nicely anymore.

Changed to let buttons grow to 100% before they break to 2 lines,
also changed the spacing a bit so that even if there are 2 lines
the icon is positioned nicely.