frontend: show spinner after unlock, before accounts are loaded
After unlock, especially the first time, it can take many seconds for
the initial set of accounts to be loaded. Until then, the app shows
nothing in the sidebar and 'Please connect your device to get
started'.

The user is easily confused, it seems as something went wrong.

We can probably speed up this loading with firmware and app changes,
but this is a quick improvement that shows a spinner until the
accounts become available.