frontend: fix issue with weekly chart when hourly data is missing
This commit addresses an issue where the weekly chart would crash
if hourly data was missing. Previously, all three filters
(week/month/year) relied on the existence of daily data to display
or not. However, the week filter uses hourly data. So to determine
whether to display the weekly chart, a new attribute,
disableWeeklyFilters, has been introduced to disable only the
 weekly chart display if hourly data is unavailable.