frontend: improve chart percentage calculation
This commit fixes a regression that was introduced in commit
a9a26db that prepended a zero to the data set. With that change
the percentage calculation was not correct anymore, leading
sometimes to infinite loading skeleton. With this update, the
percentage change is correctly computed based on the first
non-zero data point.

Additionally, this update modifies the rendering of the percentage
change. if `hasDifference` is false, it won't be rendered at
all.

Refactor variable name 'hasDifferenece' to 'hasDifference'
to correct a typo.