backend/lightning: make ligthning SDK connection synchronous
The previous asynchronous connection made difficult to properly handle
errors during the activation and also caused the redirect to the
lightning account to happen before the sdk completed connecting. This
was causing an error being shown for a few seconds right after the
activation.

The activation flow may now take a few seconds to complete, but the
frontend shows a nice loader so that should not be a big issue.