frontend/lightning: add fiat conversions to send/receive flows
This also adds a getBTCSatsAmount backend endpoint that allows to
convert a sat amount into an Amount object (FE: IAmount, BE:
FormattedAmount) with the fiat conversions. This is not the best
solution, as it would be much better to heavily refactor the invoices
parsing, removing all the FE types that depends on breez and making the
enpoints return a simpler struct containing only the used data (including
the fiat conversions).

This is a workaround that allows to show fiat conversions in the
frontend while we take our time to decide how to refactor this part.