qt: delete custom libserver.h, use auto-generated one
Not sure why we copied the contents of libserver.h initially, maybe we
needed to modify the declarations a bit. Anyway, it seems to work by
directly using the generated file.

This makes sure the changes to CGO in upgraded versions (the preamble)
are adopted automaticaly, and adding/changing functions in server.go
works automatically.