frontend/conversions: show BTC as active currency in sats mode
In BTC accounts with sats mode enabled, displaying two amounts in
sats is redundant. This commit addresses this by ensuring that when
sats mode is enabled for BTC accounts, the secondary currency is
displayed as BTC. Furthermore, BTC is now displayed in this setting
as the secondary currency on the send page and during the
transaction review process.