frontend: change lightning receive workflow
This fixes a bug that sometimes showed a qr placeholder instead of
a real code. This happend if the user edited an invoice and changed
amount to invalid amount for example empty amount.

Also added new invoice button on invoice created and success step
and slightly changed the layout to better fit on mobile screen.