frontend: reduce container spacing on mobile
With this ViewBottons are nicely positioned on mobile.

One change affects large screen that fixes a small misalingment of
aopp success buttons.

Buttons container on large screen should have no padding if inside
a View component with fitContent prop, else they are misaligned
with the content.

Button padding is only needed for small screen and defined later
in this change.