frontend: add device password recommendation
Adds a small note below the password demo video which links to the
knowledge base entry on how to choose a good device password.

Some users are confused when it comes to the device password and may
choose overly long device password. The knowledge base entry provides
valuable context and enables the user to make an informed decision.

To avoid confusion with BIP-39 passphrases, "password" has been
rephrased to "device password".