android: add com.google.android.gms.version to AndroidManifest.xml
A user reported a crash when accessing the camera for QR code
scanning. The crash report said that this tag was required.

I could not find much info about why accessing the camera would
require Google Play Services. It seems however that simply adding the
tag should not prevent the app to be installed on devices without
Google Play Services.