backend: handle connection error during lightning activation
If an error occurs during the lightning connection process, it is returned, but
the lightning configuration is not cleared. This leads to the appearance of a
lightning wallet in the app when, in reality, it doesn't exist. This commit
addresses this issue by deactivating the lightning configuration
if any error occurs during this step.