backend/btc: use mempool.space as default fee estimator
Bitcoin Core fee estimation is not very accurate. With this change,
meempool.space api will be used as a preferred estimation method for BTC
fees. The fee estimations are fetched through a Shift mirror server.

Also, the fees are not cached and updated for each new block, but they
are fetched every time they are needed, since now they depends on the
mempool state.

Duplicated fees were filtered and not sent to the backend, this filter
has been removed since it was not optimal for the UX.