frontend: change support links to redirects
Replacing external links to our knowledge base with redirects on the
website enables easy and swift updates for everyone if the knowledge
base URL ever changes in the future, e.g. because of a domain change.

This will avoid broken links in old BitBoxApp versions. The new
redirect links are already implemented on the website.