frontend: trying responsive font size
This is a test to see if we can use responsive font size. This also
sets the base font size so that 1rem is 10px (unless text zoom).

With this technique elements can have responive dimension so that
sizes using rem are also zoomed if the text font size is bigger.