frontend: andoird font size zoom can break the layout
Increasing the font size setting under 'Display size and text' on
Android can break the layout in the WebView. There are 2 relevant
settings: 'Font size' and 'Display size'.

It appears that only the text is increased leaving the elements as
if the font size was normal. Buttons look broken with text hanging
out and transaction history is unreadable with overlapping text.

This change seems to have no impact and was for testing purpose.