frontend/android: check if webview can go back or close the app
This should replace native Android back behavior with either going
back in the webview or closing the app.

This is using the native goBack method but it seems that the app
only goes back to account summary. This could probably be improved
later and seems already useful with just going back to summary.

On the account summary it cannot go back any further so it will
trigger the old do-you-want-to-quit alert.

Replaced deprecated onBackPressed override.