frontend: simpler coin summary
Coin balance needs a list of coinCodes and some way to display the
coin name.

Before this change it first constructed a map that contains accounts
per coin, as the name of the coin is only included in IAccount
interface.

Using a dedicated utils function to derive the coin name from coin
codes simplifies CoinBalance component.