frontend: type coincode with coin and tokens and keys of various maps
Some maps just typed their key just as string, this replaces some
keys with the actual types.

Before we had to typecast in a few places as CoinCode did not include
tokens but was used for some apis that also accept tokens.

Changed CoinCode to be intersection of NativeCode and TokenCode.

    type CoinCode = NativeCoinCode | ERC20CoinCode;

There are a few special tokens and coins such as: erc20Test or rbtc.