frontend: set data source to hourly on mount if display is week
Our chart uses `daily` as its `source` default state value.
Instead, we need to make sure to set this `source` state to
the appropriate value according to the stored `chartDisplay`
in the context.

Other than as a best practice, this fixes an edge case bug.
Such as wrong data & difference displayed when user does the
following:
1. picks USD as default currency,
2. navigate to a different page,
3.  and going back to account summary.