frontend: make reusable pill button & pill button group
Right now, we're only using such components in `filters.tsx`.
However, this will change in the future as other components may
incorporate these pill buttons for filtering / tabbing-like
functionality.

I didn't make a new variant in the button component because
we wanna keep the styling on the parent container of the buttons
(`PillButtonGroup`) as we want them to look & behave the same
(size-wise, etc) on every components we want to have them in.

So instead, I made a `PillButtonGroup` and `PillButton`
components to be used together.