frontend: migrate to latest i18next and react-i18next
i18next rewrote its types in latest versions, but defining
i18n returnObjects: true does not return the correct type when
only specifying a key to an object of translations.

This was used for guide entry component to point to an object
that contains title and text for the guide entry.

It is possible to define a CustomTypeOptions with i18next.d.ts to
globally returnObjects: true, but this breaks all regular uses of
the t function that should return a string.

https://www.i18next.com/overview/typescript#custom-type-options

Added local option for each t function that is used for guide
entries and remove the global option.

Pluralization is used for native notification of newly detected
transactions.

Other changes:
- added condition for i18n.resolvedLanguage as it can be undefined
- rewrote unknownError that optionally receives errorMessage
- remove type casting for account.syncedAddressesCount
- rewrote component that passed t function

Tested:
- only use of pluralisation (notification.newTxs) works
- feetarget customLabel_eth
- BB01 unlock.error.e<n> returns correct error message with fallback
- render correct translation with context (bb02Bootloader.success
  or bb02Bootloader.abort)
- process parameter works in context (bootloader.progress)
- keys using underscore (buy.info.disclaimer.payment.table.2_method)