frontend: replace inline-message with message component
InlineMessage component was used to show a message when Tor setting
was changed by the user. This message could be dismissed, but this
was not ideal as the app does not use the changed Tor config until
restarted.

Replaced with message component, changed success to warning.

With this change the message is not dismissable anymore.