frontend: extract address or txid into component
Create AddressOrTxID component that can be used to display transaciton
addresses or the txid in the details and the transaction itself.
Since we never display the txid in the transaction row right now the
component enforces this by returning null and logging an error in case it
is misused. TypeScript also enforces that either addresses OR txid is
passed to the component but not both. This makes the transaction.tsx
component easier to read and more concise.