frontend: improve bitcoin or crypto in translations
In some languages Bitcoin or crypto can have special ending depending
on the context, so translating crypto does not work in all cases.

This is the first step changing keys using bitcoin or crypto
context so that terms like Bitcoin or crypto can be translated
differently.

Added generic buy and receive keys with context, without context
this keys still use a coinCode placeholder. i.e.
- Buy Bitcoin
- Buy crypto
- Buy LTC / Buy {{coinCode}}

Reason to not offer different keys for each token/coin is that it
would required a different key for each coin/token.