backend/handlers: fix err handling in getExchangeBuySupported
If getExchangeBuySupported was called after an account was closed it was
returning a generic error, causing the frontend to display the generic
error popup when disconnecting the bitbox every now and then.

This makes the endpoint to return an empty list of supported exchanges
if the account is not available.