frontend: remove unused transaction style class
Remove the "first" class that was used in combination with "container"
its CSS rule does not contain any CSS declarations and thus does not do
anything. The CSS rule was empty after this commit:
5e1898b245bb162f234de0c7fd568974c4f151a2

Since the index prop on the transaction component is only used to add
this class it can also be removed.