android: load using loadUrl instead of loadDataWithBaseURL
When loading with the latter, webview.goBack() (or history.back() in
JS) go back to the index page instead of the latest entry. Not sure
why, but loadUrl is simpler anyway.

We still use BASE_URL so that it is used as the referer in Moonpay,
which is required for MoonPay to be able to load.

The reason the app still loads even though it looks like we are
loading shiftcrypto.ch is that we intercept it in
shouldInterceptRequest() and load from local assets instead.