backend/transaction: auto select highest fee for payment requests
There was a bug in the payment request implementation that didn't
correctly support mempool.space fees and resulted in a fee error
during the tx proposal on production environment.

This commit moves the fee target selection for payment requests in the
BE and fixes the bug.

It also fixes a bug in the tx proposal error handling.