frontend: experiment to let account render in offline mode
This would fix the flickering when changing accounts and could
allow the user to get a receive address in offline mode.

Unfortunatelly there are multiple issues with this:
- offline error only showing once
- receive probably does not know which address range to use and
  could lead to unwanted address-reuse
- backend seems to not play along well in offline mode but somewhat
  works

If we go this way we could add various skeletons to the account
view so that it is clear which elements are still loading.