frontend/backend: pass tx. note to sendtx
Pass the final transaction note to the sendtx handler on submit (send)
instead of remembering a proposed transaction note which is set
separately using the propose-tx-note handler.

Currently there is no benefit of having a separate handler for the
transaction note for the active proposal. It was separated in this
commit because typing in the note field would cause the other fields to
be validated/calculated: 4eb692c84351dbd23ae391e99ffd5d8452db08f7

However since the backend does not perform any calculations or
validation on this handler, the final note can be submitted when
clicking send (calling the sendtx handler).