frontend: don't disable back-button in send
Don't disable the back button in send.tsx when a dialog/modal (QR
scanner, or UTXO selection) is open. It is a UX best practice to do this
when a modal is open **however** a) our dialog CSS creates an overlay
which means nothing behind it can be clicked anyways, and b) the back
button currently supports ESC which should close the modal but navigates
back from the send route right now.

If we don't need the is-open state for these modals we can simplify the
send component in another refactor by extracting coin control into its
own sub-component (moving `activeCoinControl` there) and moving
`activeScanQR` into `ReceiverAddressInput`.