frontend: move qr scanner open state to sub-comp.
Move the `activeScanQR` state used to control the QR scanner dialog
(open/closed) into the `ReceiverAddressInput` to simplify the component.

We can remove the `activeScanQR` from the back button disabled property
because a) navigation with ESC allows the user to go back anyways and b)
the back button itself is not clickable when there is an active overlay.