frontend: create sub-component for coin control
Create a sub-component for the coin control toggle and dialog (utxo
selection) to make the send.tsx component smaller and easier to
understand.

We can remove `activeCoinControl` from the back button disable property
because a) the use can already navigate back using ESC, and b) the back
button is not clickable while there is an active overlay.