ci: refactor macOS build
We want to call `make ios` in another step, but using `github-ci.sh`
again for this like for Qt would repeat all the steps related to
installing Brew, setting up Go, coping the repo to $GOPATH, etc.

This commit refactors it so these steps are done upfront in separate
steps, so we can append a `make ios` step afterwards.

The macos runner image is bumped to 14, which defaults to XCode 15. The
older version had some build errors with the iOS app. Since macos-14
is ARM based, we also need to use osx-brew.sh to install the Qt deps
for amd64.

I suspect that Go does not need to be installed from amd64 brew, the
native one can cross compile fine. Something to try later. If so, we
can add this to ci.yml to install the native ARM Go:

```

    - name: Install Go
        uses: actions/setup-go@v5
        with:
          # Take Go version to install from go.mod.
          go-version-file: 'go.mod'
      - name: Set GOPATH
        run: |
          echo "GOPATH=$(go env GOPATH)" >> $GITHUB_ENV
      - name: Copy repo to GOPATH
        # This is needed as gomobile is still unaware of go modules, so the repo must be in GOPATH
        run: |
          mkdir -p $GOPATH/$(dirname $GO_SRC_DIR)
          cp -a ${{github.workspace}} $GOPATH/$(dirname $GO_SRC_DIR)
```