frontend: remove i18n placeholder in error messages
These error messages should use context so that the whole message
can be translated for each language.

Adding a placeholder does often not work in every langauge and the
term 'buy' or 'sell' was not translatable in this case.

Removed the action placeholder for these messages, this needs to
be updated in locize as well in all langauges.