frontend/backend: remove BB01 signConfirm
backend: Remove `EventSignConfirm` as we don't want to utilize it in the frontend
anymore.
frontend: Remove `touchConfirm` from the `WaitDialog` component because it is not
important anymore. As a side effect `paired` can also be removed from
the props because it was only used to apply a different styling to
content that is now removed.

Because of this several CSS rules, i18n keys/translations and assets can
be removed too. The `signConfirm` subscription in `devicessync.ts` is no
longer used and was also removed.