ios: enable software keystore in testnet
The iOS app currently can't connect to a BitBox. In the testnet app
target, we enable the software keystore for now, which we can use to
load accounts to test the app already now.

This is supposed to be temporary - we might enable another way to load
the software keystore in general later, but should disable it for
general use once a BitBox is able to connect.