ios: allow app to access camera
Needed for QR code scanning.

WebView in addition has a permission request that can answered with
"prompt" or "grant". We grant by default, as the permission was
already granted at install time, and our WebView is not a general
browser where giving permission on a per-page basis is that useful.

We could consider turning it into a prompt when the request comes from
an exchange widget in the future.