ios: fix exporting files (CSV exports, log export, notes export)
Similar to Android, we set `ExportsDir()` to `""` and let
the backend environment `getSaveFilename()` construct the full path.

Mobile handling of the dirs/files is a big hack and works on implicit
assumptions, e.g. that `getSaveFilename()` is always called with
`filepath.Join(exportsDir, filename)`, which is only the filename on
mobile as `ExportsDir()` returns `""`. This should be cleaned up, but
for now iOS is made to behave like Android in this regard.

We use a temp directory for these files, so the files will be cleaned
up automatically after a while.