frontend: fix layout on mobile for spinner with error message
Some error messages are only partly centered and have no padding
to the edge of the container.

Reason:

Spinner is used in account and other components to show loading but
also unknown error messages. Each line of the message is rendered
in its own paragraph element. Those paragraphs are centered by the
parent element using flexbox. As soon as a paragraph contains text
that spawns over multiple lines it will cause the paragraph to use
the full available width with left aligned text.