frontend: improve send error handling
The `txProposalErrorHandling` function should **only** return errors and
not set any state. Although this isn't causing an issue right now, the
function should not return an object that is unrelated to errors in
principle. Additionally, this will be necessary for an upcoming refactor
of the `Send` component.

Return only `TProposalError` from the function but make all fields
optional. To keep the logic the same we set `proposedFee` to `undefined`
according to the same logic (ie. `{}` or `{ amountError: ... }` is
returned) but in the `Send` component not in the error handling
function.