frontend: improve send error handling
The `txProposalErrorHandling` function should **only** return errors and
not set any state. Return only `TProposalError` from the function but
make all fields optional. Although this isn't causing an issue right
now, the function should not set state in principle. Additionally, this
will be necesary for an upcoming refactor of the `Send` component.

To keep the logic the same we set `proposedFee` to `undefined` according
to the same logic (ie. `{}` or `{ amountError: ... }` is returned) but
in the `Send` component not in the error handling function.