frontend: improve proposal response state
Instead of individual variables `proposedFee`, `proposedTotal`, and
`proposedAmount` we can put the result itself into the component state.

Because of this, it is not possible anymore to reset individual fields
to `undefined` but this is correct since there should be no "old"
proposal values available in case the current proposal was invalid
resulting in an `undefined` `proposalResult`.

Before this commit the amount fields could show false values because
they would show the last converted `fiatAmount` and the last valid
`proposedAmount` since it was not reset on error. Now that we reset it
it would've shown `''` but this is not a good solution which is why this
commit also falls back to the last typed amount and its fiat equivalent
in case `sendAll` is `true` but the proposal was invalid.

We also disable the `sendAll` checkbox in case there is an
`addressError` or `feeError` because it can not work if the proposal is
not valid. We don't disable it if there's only `amountError` because the
`amount` is not relevant if `sendAll` is `true`.