frontend: don't reset the custom fee
Don't reset the custom fee when the fee target changes/is set. The
backend only uses the custom fee if fee target = custom which means it
does not need to be set to be empty ('').

Like this the user can select custom, type something, change to some
other target, but as soon as he chooses custom again, what he typed
earlier is still there.

This will also enable us to only pass a `setFeeTarget` to the
sub-component instead of having to define a handler, later when we
refactor the `Send` component to functional.