frontend: update 'My portfolio' with Lightning account
This commit adds functionality to display Lightning account
in 'My portfolio' in cases where it is active but not associated
with a connected or remembered wallet, or when all mainnet
accounts in the connected wallet are disabled.

The total coins table is updated to include the Lightning account
Hide chart, percentage change and filters in my portfolio if
only lightning account is displayed.