frontend/qrcodescanner: improve qrscanner component
7c0b2f6 introduced a 300ms wait between the QrScanner creation and its
start to fix a bug on re-renders of the component, but it seems not
to be enough on certain devices, which still show the bug.

It would be nice to change the 300ms fixed wait to a better way to know
when the scanner is ready to be started but we could not find it.

This commit improves the general structure of the component, to limit
race conditions across re-renders and seems to improve the behavior and
fix the bug with the tested devices.