frontend: fix lineheight and fontsize of empty wallet message
Line-height of the empty wallet message was 1 and looked squeezed.

Changed to SubTitle component, but the font-size used for this
was smaller than paragraph. Changed SubTitle on large screen so
that it has same font-size as paragraph. Could even be bigger in
the future.

SubTitle component is only used in Account and empty account message
and in General Settings.