qt: force mime type resolution by extension
Fixes #2684 and hopefully #3061.

Without this, on linux the local mimetype database (influenced by
various packages installed on the system) can mess with the mimetype
resolution of our .html/.js/.css files served locally, resulting in a
blank page.

This manually intercepts local requests and forces the correct
mimetype for a set of hardcoded extensions.