ci: build windows app and report artifact
This adds a GitHub Action job to build the Windows app and create the
installer. Artifacts will be reported on Mattermost, like with other
platforms.

Notable differences to our previous Appveyor CI:

* The GitHub windows-2019 image comes with Vistual Studio Enterprise
instead of the Community edition. A directoy junction is created to
point calls from the Community to the Enterprise path.
* NSIS, MinGW and Make are pre-installed already, not requiring any
installations with choco anymore.
* Qt is not pre-installed anymore and now installed using an Action.

Because there are no changes to other files, the Appveyor build will
continue to work and can be used in parallel.