ios: add initial bluetooth functionality
In this initial prototype, the app connects to the bluetooth dev-kit
and wires up communication so the BitBoxApp can talk to it as if it
was a BitBox02. In the future we want to enumerate BitBoxes in the UI,
but for now we simply connect to the devkit immediately.

The backend environment DeviceInfo function is used by the backend to
recognize if a device is connected. We use it to recognize a bluetooth
device. The interface is geared for USB HID devices, so for now we
simply pretend to be a USB device. In a future PR we should change the
DeviceInfo to be more generic, or have a separate method for Bluetooth
devices.