frontend: revamp buysell action buttons
Revamp buy sell action buttons. Previously we use 2-step approach
here. First user needs to select via radio button, then click on
the action button (next).

In this commit, we changed the radio buttons to become
clickable card / normal  buttons called ActionableItem.
And when clicked, this will bring the user to the next screen.
This eliminates the need for the next button and shortens the
steps the user need to take.

On screen width <=400px, we hide the exchange badges to save
space and also set the back button to full width