backend/bitbox02: add some simulator tests
keystore.go, especially the BTC sign function, is untested and quite
long/complicated.

bitbox02-api-go has simulator tests, for reference:
https://github.com/BitBoxSwiss/bitbox02-api-go/blob/0e6a62c1fcf54ebaace25f90f1ab67530fedf171/api/firmware/device_test.go#L165.

If the keystore signing code lived in the bitbox02-api-go repo, we
could more easily test it, and potential third parties would also have
an easier time integrating the BitBox.

As a first step, we add simulator tests to the bitbox-wallet-app repo
for bitbox02/keystore.go, which would enable us to safely refactor the
code so we can eventually move it to bitbox02-api-go.

The code to download and run simulators was copied from
bitbox02-api-go:
https://github.com/BitBoxSwiss/bitbox02-api-go/blob/0e6a62c1fcf54ebaace25f90f1ab67530fedf171/api/firmware/device_test.go#L48