backend: allow opening checkout.btcdirect.eu
Some payment providers such as iDEAL or SOFORT need to breakout
the BitBoxApp and be continued in a native browser.

From https://checkout.btcdirect.eu/ the user fills in the bank they
are using and will get redirected to their banking environment.

On mobile the bank may use intent links to automatically open and
continue in the banking app.