frontend: always show manage-device in settings menu
Conditionally hiding navigation items is not ideal and can be
confusing.

Another reason to always show 'Manage device' but with 'No device
connected' is to have a place to add detect-bitbox button.

With this change the app will route into device settings if the
user unlocks on no-connected-device route.

It does not cover showing no-connected-device again after the user
unplugs, reason is that with rememeber-wallets this should go to
account-summary and without go to waiting for device. There was
no clean way to detect and route back to no-connected-device.

Ideally the 'Manage accounts' item is also always shown in settings
navigation, which can be done in another commit.