backend: remove blocking API calls.
We don't want calls to Balance, Transactions,
GetUnusedReceiveAddressesses and SpendableOutputs to block as this
can cause calls to pile up and consume resources; also, this calls
can be made irrelevant if the user switches to a different page.

Instead, while the first sync is in progress, we return an
appropriate error. After that, every call returns the latest known
state.