qt: fix uri scheme handling on macOS
In Qt 6.8, the way custom uri schemes are handled on macOS changed.

The commit that changed it is
https://github.com/qt/qtbase/commit/664c7ffb212eb898ed03f7b19c883400fa027b6b. Before,
the URL would be emitted as a FileOpen event with a QFileOpenEvent
object. After this change, it is dispatched to QDesktopServices.

Before, QDesktopServices would only be used to register custom schemes
while the app was running, but now it also is used when the app is
launched with by a URI click (provided the Info.plist entry for it is there
of course).