ios: show bluetooth picker in Manage Device
When there are watchonly accounts, the waiting screen is not
shown. Instead, the user can connect a new device in Manage Device.

Manage accounts also shows the same screen if there are no
accounts. The assumption is that if there are no accounts, then there
is no device connected, so we might as well show the device picker
there too.