frontend: integrate bitrefill widget
Integrates Bitrefill in the frontend. On the previous "Buy & sell"
menu, there is now a third selectable pill "Spend", opening the
available exchange options (Bitrefill in this case).

The widget is requested from bitboxapp.shiftcrypto.io using config
parameters from the user/app, handled in the "request-configuration"
event. Note that the iframe will still have its source set to
"https://embed.bitrefill.com".

When the user clicks "Pay" inside the Bitrefill widget to purchase
a gift card, the returned address and amount is used to propose a
transaction on the connected BitBox.

The PaymentMethod component is adjusted to allow custom messages
to be shown in the exchange overview, as these will differ in the
"Spend" section between providers.

Disclaimer and info texts are added technically, but the actual
text is still missing in this commit.

Bitrefill developer documentation:
https://www.bitrefill.com/playground/documentation/url-params