backend/keystore: use accountConfig&derivation in VerifyAddressBTC
A BTC address is derived from an account using
change/addressIndex. There is no "address-level" signing
configuration. `signingConfig.Derive()` assumes there is one base
level keypath and xpub, which is not true for a descriptor in general. For
example, a multisig descriptor has multiple xpubs, each with their own keypath.