backend/btc: remove address-level signing config
The previous commits moved towards using an account config plus
derivation info to get to the address. This commit removes the
address-level signing config for BTC.

This is a simplification, as it does not make sense to derive a full "signing config" when all
that is derived is the keypath, nothing else changes.

This is part of moving to general descriptor support. There are still
remaining issues where code assumes it's a single-sig (calls to
address.AbsoluteKeypath(), and NewAddress(), etc.), but we can solve
that later step by step.