frontend: show all sell options
Before BTC Direct sell the only sell option was Bitcoin via
PocketBitcoin, which require SLIP24 and only supports Bitcoin.

If the firmware did not support SLIP24 the app displayed:

    A firmware update is required to make use of this feature.

When an altcoin was selected the app would display:

    An account must be BTC based to support payment requests.

This commit changes to show all sell options and optionally
show an warning after the options:

    A firmware update is required to use all options.

This requires a change in the backend that the Pocket options is
only returned if the firmware supports SLIP24.