qt: use persisted profile, not off-the-record
When we upgraded from Qt 5.15 to Qt 6, the default profile seemed to
have changed from being stored on disk to off-the-record:

- https://doc.qt.io/qt-5/qwebengineprofile.html#defaultProfile
- https://doc.qt.io/qt-6.5/qwebengineprofile.html#defaultProfile

This made the third party widgets (MoonPay, Pocket, ...) lose their
cookies/localstorage upon app restart.

This commit switches our profile to be stored on disk once more. We
choose a new storage name "BitBoxApp" (not "Default" which was used
previously) to not run into potential conflicts or very old
user sessions that get unexpectedly resumed.