ios: activate scanning only when device picker is shown
Currently, e.g. in watchonly, peripheral scanning would be active all
the time, draining the phone battery.

With this commit, scanning is started when the device picker is
mounted (visible) and stopped when unmounted (not visible).

Special case: Bluetooth is turned off after the device picker is
mounted. In this case, we remember that we want to be scanning, and
start doing so when Bluetooth is turned on.