exchange: refactor backend and frontend code for marketplace
Previous code had a structure that naturally allowed to buy and sell
coins. The recently introduced spend action, that allowed to start
integrating Bitrefill, slightly changed some of the main concepts of
the implementation,e.g Exchanges and Deals.

This refactor converts the whole exchange section into Market, allowing
a more natural addition of new services, not limited to buying and
selling.

Exchange partners and other providers are now identified as market
Vendors, while the concept of Deal remains, but will be made more
flexible in future improvements.