logging: remove unused logging.json config file
It stored the path to log.txt and output level. Not sure why, it is
not used for anything really, it was always log.txt in the app
folder (which the log export etc assume).

That this file was stored in the app folder caused an issue on iOS:
the ID of the app can change especially when running debug builds in
xcode, but the logging.json in the app folder of a previous debug run
would contain a log.txt filepath that became invalid.