fronted/qt: fix web engine profile for external links
In 00d7607 we switched from the default profile to a dedicated one, to
store profile data on disk. This change didn't involve the WebEnginePage
used to display external resources (e.g. links in 3rd party widgets) and
caused external links not to be displayed because of incompatible
profiles.

This fixes the issue sharing the same profile also for external pages.