backend: close Electrum/failover connections too on shutdown
Without this, if no backends/servers are configured, headers download
would be stuck on some call that keeps failing over or retrying
forever, and headers.Close() would be blocked from finishing.

This is a quick fix, there is still a delay to shut down until some
Electrum call in headers.go fails (probably up to 30s due to the retry timeout?).

More things to do:

- Add a ctx var for every Electrum call that should get cancelled on
shutdown to not delay
- Do not block closing headers on the same lock which is held during a
blockchain call (use a different lock, or by the above ctx solution)