chore: remove pointer receiver from AccountsConfig.
In order to have a single type of receiver for the struct, we need to
either have every method with a pointer receiver, or a value one. A
value one is not possible since GetOrAddKeystore needs to modify the
config.
A pointer one is also troublesome because AccountsConfig() returns a
value, so calls like b.AccountsConfig().Derive() or similar wouldn't be
possible.
At the same time AccountsConfig() can't be changed to return a pointer
because it needs to lock the accountsConfigLock and returning a pointer
would cause data races.