frontend: fix regression dialog should conditionally call onclose
Fixed an issue with FactoryResetSetting where continuing to reset
the device should show a WaitDialog.

The simplified Dialog version after refactoring always called
onClose as if the user would aboard the workflow.

Instead it should only call onClose if the user closed the dialog
(close button, esc or clicking on the background) but NOT if the
parent closed the Dialog.

Regression introduced in:
- d5bb07888deac147a74b527c0251529f6ef0174f