frontend: deprecate --size-wizard-text variable
Was only used during setup for the label of the name input field
and the backup checklist for the checkboxes.

In both views there is no reason to use another font-size, without
it the text looks more consistent and label, input field and error
message have the same font-size as in other palces.