frontend/android: refactor network related code
Our MainActivity code is quite complicated and mixes a lot of different
things. This change moves the code related to network management (e.g. to
display alerts when the network is not available) to a dedicated
helper class and declutters the main activity a bit.

It also fixes a bug that caused the app to misdetect mobile connection
or missing connectivity when switching between WIFI and cellular
connection. The bug was caused by the fact that the "onLost" call is not
guaranteed to be executed after the network is actually lost. This can
bring both to false positive and false negatives results, depending on
race conditions. This caused to occasionally not display the "connection
lost" banner when needed, or to e.g. display both the "connection lost" and
the "mobile connection" banner at the same time.

Adding a 250ms delay in the execution of the "onLost" handler is not
nice, but seems to reliably fix the issue on all the tested Android
versions.

This also drops the `netWorkStateReceiver`, as it was registered as a
broadcast receiver for the `CONNECTIVITY_ACTION`, which is now
deprecated.
See https://developer.android.com/reference/android/net/ConnectivityManager#CONNECTIVITY_ACTION

The `usingMobileDataChanged` call, which is used to force the mobile
server to fetch again the network source (wifi/mobile) and optionally
display the "mobile connection" banner, is moved inside the
`checkNetworkConnectivity` call, so that it's executed every time we are
reassessing the connectivity status.