android: align system bars with app theme
The colors of both the status and the navigation bars and their icons
were not always matching the app background.

Also, forcing the app theme in the settings and then switching the OS
theme was sometime causing the bar icons to change and blend with the
bars.

This commit does the following:
- Drives the colors of both bars and icons in the setDarkTheme method
to avoid misalignemnts with the app theme
- Removes the onConfigurationChanged overriding, which was conficting
with the implemented theme handling, which is delegated to the frontend
DarkmodeProvider. Listening of OS theme changes in the activity must be
avoided as it could fire unwanted changes in the app theme.
- Improve the app static styles, to reduce the color flickering before
setDarkTheme() is called for the first time.
- Apply matching backgrounds to the root layout and WebView so the bars
blend with the app in both light and dark modes.