frontend: fix wrong outgoing amount regression
In the current transaction overview the app shows what amount goees
out of the wallet. In case of a send2self transaction only the fee
is outgoing, but hte amount is sent to self.

This was wrongly changed in commit:
- d5ef18f8fc8ff5759e0e70e26115f06c2b6cc867

Fixed by partly restoring the old logic, but still keeping the new
AmountWithUnit component.